+++ character/mobile/character.js
@@ -19,7 +19,6 @@ const characters = {
 	jiangji: ["male", "wei", 3, ["twjichou", "jilun"], ["character:tw_jiangji", "die:tw_jiangji"]],
 	liwei: ["male", "shu", 4, ["jiaohua"], ["die:tw_liwei"]],
 	laimin: ["male", "shu", 3, ["laishou", "luanqun"]],
-	yj_zhoubuyi: ["male", "wei", 3, ["mbhuiyao", "mbquesong"]],
 	xin_guozhao: ["female", "wei", 3, ["yichong", "wufei"]],
 	xin_zhangyi: ["male", "shu", 4, ["xinwurong", "shizhi"]],
 	xin_sunliang: ["male", "wu", 3, ["xinzhizheng", "xinkuizhu", "xinlijun"], ["zhu"]],
@@ -30,7 +29,6 @@ const characters = {
 	re_yanwen: ["male", "qun", 4, ["reshuangxiong"], ["name:颜|良-文|丑"]],
 	xin_zhoutai: ["male", "wu", 4, ["buqu", "new_fenji"]],
 	re_caozhi: ["male", "wei", 3, ["reluoying", "rejiushi", "chengzhang"]],
-	yj_weiyan: ["male", "qun", "4/4/1", ["mbguli", "mbaosi"]],
 	re_chenqun: ["male", "wei", 3, ["redingpin", "refaen"]],
 	xin_caoxiu: ["male", "wei", 4, ["qianju", "xinqingxi"]],
 	xin_zhuhuan: ["male", "wu", 4, ["fenli", "xinpingkou"]],
@@ -55,7 +53,6 @@ const characters = {
 	re_wuguotai: ["female", "wu", 3, ["reganlu", "buyi"], ["name:丁|null"]],
 	xin_sunxiu: ["male", "wu", 3, ["mobileyanzhu", "mobilexingxue", "zhaofu"], ["zhu"]],
 	sp_maojie: ["male", "wei", 3, ["bingqing", "yingfeng"]],
-	yj_huangzhong: ["male", "qun", 4, ["spshidi", "spyishi", "spqishe"]],
 	re_yufan: ["male", "wu", 3, ["zhiyan", "rezongxuan"]],
 	sunhanhua: ["female", "wu", 3, ["chongxu", "miaojian", "shhlianhua"]],
 	yanpu: ["male", "qun", 3, ["huantu", "bihuo"]],
@@ -96,10 +93,6 @@ const characters = {
 	re_dongzhuo: ["male", "qun", 8, ["rejiuchi", "roulin", "benghuai", "baonue"], ["zhu"]],
 	re_xusheng: ["male", "wu", 4, ["repojun"]],
 	sp_sufei: ["male", "qun", 4, ["zhengjian", "gaoyuan"]],
-	yj_zhangliao: ["male", "qun", 4, ["weifeng"]],
-	yj_zhanghe: ["male", "qun", 4, ["xinzhilve"]],
-	yj_xuhuang: ["male", "qun", 4, ["xinxhzhiyan"]],
-	yj_ganning: ["male", "qun", 4, ["gnjinfan", "gnsheque"]],
 	re_sunjian: ["male", "wu", 4, ["gzyinghun", "repolu"]],
 	zhengxuan: ["male", "qun", 3, ["zhengjing"]],
 	dengzhi: ["male", "shu", 3, ["jimeng", "shuaiyan"]],
