+++ character/mobile/skill.js
@@ -3026,209 +3026,6 @@ const skills = {
 			},
 		},
 	},
-	//☆周不疑
-	mbhuiyao: {
-		audio: 2,
-		enable: "phaseUse",
-		usable: 1,
-		// filterTarget:lib.filter.notMe,
-		content: function () {
-			"step 0";
-			player.damage("nosource");
-			"step 1";
-			if (game.countPlayer() < 2) event.finish();
-			if (game.countPlayer() == 2)
-				event._result = {
-					bool: true,
-					targets: [game.findPlayer(i => i != player), player],
-				};
-			else
-				player
-					.chooseTarget(
-						`慧夭：请选择两名角色`,
-						`令不为你的第一名角色视为对第二名角色造成过1点伤害。`,
-						(card, player, target) => {
-							if (!ui.selected.targets.length) return player != target;
-							return true;
-						},
-						2,
-						true
-					)
-					.set("multitarget", true)
-					.set("targetprompt", ["伤害来源", "受伤角色"])
-					.set("ai", target => {
-						return target == get.event("aiTargets")[ui.selected.targets.length] ? 10 : 0;
-					})
-					.set("aiTargets", lib.skill.mbhuiyao.getUnrealDamageTargets(player, [game.filterPlayer(i => i != player), game.filterPlayer()], true));
-			"step 2";
-			if (result.bool) {
-				var targets = result.targets;
-				player.line2(targets, "green");
-				game.delaye();
-				targets[1].damage(targets[0], "unreal");
-			}
-		},
-		getUnrealDamageTargets: (player, lists, forced) => {
-			const targets = [null, null];
-			let sourceList, targetList;
-			if (lists.length == 2 && lists.every(l => Array.isArray(l))) {
-				sourceList = lists[0];
-				targetList = lists[1];
-			} else {
-				sourceList = lists.slice();
-				targetList = lists.slice();
-			}
-			const list = targetList
-				.map(current => {
-					const _hp = current.hp,
-						_maxhp = current.maxHp;
-					current.hp = 100;
-					current.maxHp = 100;
-					const att = -get.sgnAttitude(player, current);
-					let val = get.damageEffect(current, player, current) * att;
-					current.getSkills(null, false, false).forEach(skill => {
-						const info = get.info(skill);
-						if (info && info.ai && (info.ai.maixie || info.ai.maixie_hp || info.ai.maixie_defend)) val = Math[val > 0 ? "max" : "min"](val > 0 ? 0.1 : -0.1, val + 2 * att);
-					});
-					const eff = 100 / val + 15;
-					current.hp = _hp;
-					current.maxHp = _maxhp;
-					return [current, eff];
-				})
-				.sort((a, b) => b[1] - a[1])[0];
-			if (list[1] < 0 && !forced) return targets;
-			const targetx = list[0];
-			targets[1] = targetx;
-			const list2 = sourceList
-				.filter(i => i != targetx)
-				.map(current => {
-					const _hp = targetx.hp,
-						_maxhp = targetx.maxHp;
-					targetx.hp = 100;
-					targetx.maxHp = 100;
-					const att = -get.sgnAttitude(player, current);
-					const eff = get.damageEffect(targetx, current, current) * att;
-					targetx.hp = _hp;
-					targetx.maxHp = _maxhp;
-					return [current, eff];
-				})
-				.sort((a, b) => b[1] - a[1])[0];
-			if (!list2) return targets;
-			targets[0] = list2[0];
-			return targets;
-		},
-		ai: {
-			order: 6,
-			result: {
-				player: function (player) {
-					if (player.getHp() + player.countCards("hs", card => player.canSaveCard(card, player)) <= 1) return 0;
-					var limit = 25;
-					var quesong = player.hasSkill("mbquesong") && !player.getStat().damaged;
-					if (quesong) {
-						limit -= 7.5;
-					}
-					if (
-						quesong &&
-						game.hasPlayer(target => {
-							var att = get.attitude(player, target);
-							if (att < 0) return false;
-							return (
-								att *
-									Math.sqrt(
-										Math.max(
-											1,
-											[1, 2, 3, 4].reduce((p, c) => p + target.countEmptySlot(c), 0)
-										)
-									) >=
-									10 || target.getHp() <= 2
-							);
-						})
-					)
-						return 1;
-					if (
-						!quesong &&
-						game.hasPlayer(target => {
-							if (target == player) return false;
-							var _hp = target.hp,
-								_maxhp = target.maxHp;
-							target.hp = 100;
-							target.maxHp = 100;
-							var att = -get.sgnAttitude(player, target);
-							var val = get.damageEffect(target, player, target) * att;
-							target.getSkills(null, false, false).forEach(skill => {
-								var info = get.info(skill);
-								if (info && info.ai && (info.ai.maixie || info.ai.maixie_hp || info.ai.maixie_defend)) val = Math[val > 0 ? "max" : "min"](val > 0 ? 0.1 : -0.1, val + 2 * att);
-							});
-							var eff = 100 / val;
-							target.hp = _hp;
-							target.maxHp = _maxhp;
-							if (eff < limit) return false;
-							return true;
-						})
-					)
-						return 1;
-					return 0;
-				},
-			},
-		},
-	},
-	mbquesong: {
-		audio: 2,
-		trigger: { global: "phaseJieshuBegin" },
-		filter: function (event, player) {
-			return player.getHistory("damage").length;
-		},
-		direct: true,
-		content: function () {
-			"step 0";
-			player.chooseTarget(get.prompt2("mbquesong")).set("ai", target => {
-				var player = _status.event.player;
-				if (get.attitude(player, target) <= 0) return 0;
-				var len = lib.skill.mbquesong.getNum(target),
-					hp = target.getHp();
-				return len + target.isTurnedOver() * 2 + (1.5 * Math.min(4, target.getDamagedHp())) / (hp + 1);
-			});
-			"step 1";
-			if (result.bool) {
-				var target = result.targets[0];
-				event.target = target;
-				player.logSkill("mbquesong", target);
-				var len = lib.skill.mbquesong.getNum(target);
-				if (target.isHealthy()) event._result = { index: 0 };
-				else {
-					target
-						.chooseControl()
-						.set("choiceList", ["摸" + get.cnNumber(len) + "张牌并复原武将牌", "回复1点体力"])
-						.set("prompt", "雀颂：请选择一项")
-						.set("ai", () => {
-							var player = _status.event.player;
-							var len = _status.event.len;
-							return get.effect(player, { name: "draw" }, player, player) * len >= get.recoverEffect(player, player, player) ? 0 : 1;
-						})
-						.set("len", len);
-				}
-			} else event.finish();
-			"step 2";
-			if (result.index == 1) {
-				target.recover();
-				event.finish();
-			} else target.draw(lib.skill.mbquesong.getNum(target));
-			"step 3";
-			target.link(false);
-			"step 4";
-			target.turnOver(false);
-		},
-		getNum: function (player) {
-			return player.countCards("e", card => get.subtype(card) != "equip5") >= 3 ? 2 : 3;
-		},
-		ai: {
-			expose: 0.2,
-			maixie: true,
-			skillTagFilter: function (player, tag) {
-				if (player.getStat().damaged) return false;
-			},
-		},
-	},
 	//郭照
 	yichong: {
 		initSkill: function (skill) {
@@ -5128,110 +4925,6 @@ const skills = {
 		},
 		ai: {
 			expose: 0.2,
-		},
-	},
-	//☆胃炎
-	mbguli: {
-		audio: 2,
-		enable: "phaseUse",
-		filterCard: true,
-		selectCard: -1,
-		position: "h",
-		filter: function (event, player) {
-			if (player.hasSkill("mbguli_used")) return false;
-			var hs = player.getCards("h");
-			if (!hs.length) return false;
-			for (var card of hs) {
-				var mod2 = game.checkMod(card, player, "unchanged", "cardEnabled2", player);
-				if (mod2 === false) return false;
-			}
-			return event.filterCard(get.autoViewAs({ name: "sha" }, hs));
-		},
-		viewAs: {
-			name: "sha",
-			storage: { mbguli: true },
-		},
-		onuse: function (links, player) {
-			player.addTempSkill("mbguli_used", "phaseUseAfter");
-		},
-		ai: {
-			order: 1,
-			threaten: 1.14,
-			unequip_ai: true,
-			skillTagFilter: function (player, tag, arg) {
-				if (arg && arg.name == "sha" && arg.card && arg.card.storage && arg.card.storage.mbguli) return true;
-				return false;
-			},
-		},
-		subSkill: {
-			used: {
-				audio: "mbguli",
-				trigger: { global: "useCardAfter" },
-				charlotte: true,
-				prompt2: "失去1点体力，然后将手牌摸至体力上限",
-				check: function (event, player) {
-					var num = player.maxHp - player.countCards("h");
-					return (num >= 3 && player.hp >= 2) || (num >= 2 && player.hp >= 3);
-				},
-				filter: function (event, player) {
-					return (
-						event.card.storage &&
-						event.card.storage.mbguli &&
-						game.hasPlayer2(current => {
-							return current.hasHistory("sourceDamage", evt => evt.card == event.card);
-						})
-					);
-				},
-				content: function () {
-					"step 0";
-					player.loseHp();
-					"step 1";
-					player.drawTo(player.maxHp);
-				},
-				group: "mbguli_unequip",
-			},
-			unequip: {
-				trigger: {
-					player: "useCardToPlayered",
-				},
-				filter: function ({ card }) {
-					return card.name == "sha" && card.storage && card.storage.mbguli;
-				},
-				forced: true,
-				popup: false,
-				logTarget: "target",
-				content: function () {
-					trigger.target.addTempSkill("qinggang2");
-					trigger.target.storage.qinggang2.add(trigger.card);
-					trigger.target.markSkill("qinggang2");
-				},
-			},
-		},
-	},
-	mbaosi: {
-		audio: 2,
-		trigger: { source: "damageSource" },
-		forced: true,
-		filter: function (event, player) {
-			return player.inRange(event.player) && player.isPhaseUsing() && event.player.isIn() && !player.getStorage("mbaosi_inf").includes(event.player);
-		},
-		logTarget: "player",
-		content: function () {
-			player.addTempSkill("mbaosi_inf", "phaseUseAfter");
-			player.markAuto("mbaosi_inf", [trigger.player]);
-		},
-		subSkill: {
-			inf: {
-				charlotte: true,
-				onremove: true,
-				forced: true,
-				intro: { content: "对$使用牌无次数限制" },
-				mod: {
-					cardUsableTarget: function (card, player, target) {
-						if (player.getStorage("mbaosi_inf").includes(target)) return true;
-					},
-				},
-			},
 		},
 	},
 	// 界曹休
@@ -7769,108 +7462,6 @@ const skills = {
 					targetInRange: () => true,
 				},
 				intro: { content: "使用牌无距离限制" },
-			},
-		},
-	},
-	//猩黄忠
-	spshidi: {
-		audio: 2,
-		trigger: { player: ["phaseZhunbeiBegin", "phaseJieshuBegin"] },
-		forced: true,
-		zhuanhuanji: "number",
-		filter: function (event, player) {
-			return player.countMark("spshidi") % 2 == ["phaseJieshu", "phaseZhunbei"].indexOf(event.name);
-		},
-		content: function () {
-			player.changeZhuanhuanji("spshidi");
-		},
-		mod: {
-			globalFrom: function (from, to, distance) {
-				if (from.countMark("spshidi") % 2 == 0) return distance - 1;
-			},
-			globalTo: function (from, to, distance) {
-				if (to.countMark("spshidi") % 2 == 1) return distance + 1;
-			},
-			aiOrder: function (player, card, num) {
-				if (player.countMark("spshidi") % 2 == 0 && card.name == "sha" && get.color(card) == "black") return num + 0.1;
-			},
-		},
-		mark: true,
-		marktext: "☯",
-		intro: {
-			content: function (storage, player) {
-				return "已转换过" + (storage || 0) + "次";
-			},
-		},
-		ai: {
-			directHit_ai: true,
-			skillTagFilter: function (player, tag, arg) {
-				if (!arg || !arg.card || !arg.target || arg.card.name != "sha") return false;
-				return player.countMark("spshidi") % 2 == 0 && get.color(arg.card) == "black";
-			},
-		},
-		group: ["spshidi_use", "spshidi_beused"],
-		subSkill: {
-			use: {
-				trigger: { player: "useCard" },
-				forced: true,
-				filter: function (event, player) {
-					return event.card.name == "sha" && player.countMark("spshidi") % 2 == 0 && get.color(event.card, false) == "black";
-				},
-				content: function () {
-					trigger.directHit.addArray(game.players);
-				},
-			},
-			beused: {
-				trigger: { target: "useCardToTargeted" },
-				forced: true,
-				filter: function (event, player) {
-					return event.card.name == "sha" && player.countMark("spshidi") % 2 == 1 && get.color(event.card, false) == "red";
-				},
-				content: function () {
-					trigger.directHit.add(player);
-				},
-			},
-		},
-	},
-	spyishi: {
-		audio: 2,
-		trigger: { source: "damageBegin2" },
-		filter: function (event, player) {
-			return player != event.player && event.player.countCards("e") > 0;
-		},
-		check: function (event, player) {
-			return (
-				get.damageEffect(event.player, player, player) <= 0 ||
-				(get.attitude(player, event.player) <= 0 &&
-					!event.player.hasSkillTag("noe") &&
-					event.player.hasCard(function (card) {
-						return get.value(card) > 9 - event.player.hp;
-					}, "e"))
-			);
-		},
-		logTarget: "player",
-		content: function () {
-			trigger.num--;
-			player.gainPlayerCard(trigger.player, "e", true);
-		},
-	},
-	spqishe: {
-		enable: "chooseToUse",
-		viewAs: { name: "jiu" },
-		filterCard: { type: "equip" },
-		position: "hes",
-		viewAsFilter: function (player) {
-			return player.hasCard({ type: "equip" }, "ehs");
-		},
-		check: function (card) {
-			if (_status.event.type == "dying") return 1 / (get.value(card) || 0.5);
-			return 5 - get.value(card);
-		},
-		locked: false,
-		mod: {
-			maxHandcard: function (player, num) {
-				return num + player.countCards("e");
 			},
 		},
 	},
@@ -17108,585 +16699,6 @@ const skills = {
 		logTarget: "player",
 		content: function () {
 			player.discardPlayerCard(trigger.player, "e", true, trigger.player.countCards("e"));
-		},
-	},
-	xinzhilve: {
-		enable: "phaseUse",
-		audio: "zhilve",
-		usable: 1,
-		chooseButton: {
-			dialog: function (event, player) {
-				var list = ["移动场上的一张牌", "摸一张牌并视为使用一张【杀】"];
-				var choiceList = ui.create.dialog("知略：失去1点体力并...", "forcebutton", "hidden");
-				choiceList.add([
-					list.map((item, i) => {
-						return [i, item];
-					}),
-					"textbutton",
-				]);
-				return choiceList;
-			},
-			filter: function (button, player) {
-				if (button.link == 0) return player.canMoveCard();
-				return player.hasUseTarget({ name: "sha", isCard: true }, false);
-			},
-			check: function (button) {
-				return button.link;
-			},
-			backup: function (links) {
-				return lib.skill["xinzhilve_" + ["move", "use"][links[0]]];
-			},
-			prompt: function () {
-				return "请选择【杀】的目标";
-			},
-		},
-		ai: {
-			order: function (item, player) {
-				return get.order({ name: "sha" }) + 0.1;
-			},
-			result: {
-				player: function (player) {
-					if (player.hp > 2 && player.hasValueTarget({ name: "sha" })) return 1;
-					return 0;
-				},
-			},
-		},
-	},
-	xinzhilve_move: {
-		audio: "zhilve",
-		filterCard: function () {
-			return false;
-		},
-		selectCard: -1,
-		delay: false,
-		content: function () {
-			"step 0";
-			event.forceDie = true;
-			if (!player.canMoveCard(null, event.nojudge)) {
-				event.finish();
-				return;
-			}
-			var next = player.chooseTarget(2, function (card, player, target) {
-				if (ui.selected.targets.length) {
-					var from = ui.selected.targets[0];
-					var js = from.getCards("j");
-					for (var i = 0; i < js.length; i++) {
-						if (_status.event.nojudge) break;
-						if (target.canAddJudge(js[i])) return true;
-					}
-					if (target.isMin()) return false;
-					var es = from.getCards("e");
-					for (var i = 0; i < es.length; i++) {
-						if (target.canEquip(es[i])) return true;
-					}
-					return false;
-				} else {
-					var range = "ej";
-					if (_status.event.nojudge) range = "e";
-					return target.countCards(range) > 0;
-				}
-			});
-			next.set("nojudge", event.nojudge || false);
-			next.set("ai", function (target) {
-				var player = _status.event.player;
-				var att = get.attitude(player, target);
-				var sgnatt = get.sgn(att);
-				if (ui.selected.targets.length == 0) {
-					if (att > 0) {
-						if (
-							!_status.event.nojudge &&
-							target.countCards("j", function (card) {
-								return game.hasPlayer(function (current) {
-									return current.canAddJudge(card) && get.attitude(player, current) < 0;
-								});
-							})
-						)
-							return 14;
-						if (
-							target.countCards("e", function (card) {
-								return (
-									get.value(card, target) < 0 &&
-									game.hasPlayer(function (current) {
-										return current != target && get.attitude(player, current) < 0 && current.canEquip(card);
-									})
-								);
-							}) > 0
-						)
-							return 9;
-					} else if (att < 0) {
-						if (
-							game.hasPlayer(function (current) {
-								if (current != target && get.attitude(player, current) > 0) {
-									var es = target.getCards("e");
-									for (var i = 0; i < es.length; i++) {
-										if (get.value(es[i], target) > 0 && current.canEquip(es[i]) && get.value(es[i], current) > 0) return true;
-									}
-								}
-							})
-						) {
-							return -att;
-						}
-					}
-					return 0;
-				}
-				var es = ui.selected.targets[0].getCards("e");
-				var i;
-				var att2 = get.sgn(get.attitude(player, ui.selected.targets[0]));
-				for (i = 0; i < es.length; i++) {
-					if (sgnatt != 0 && att2 != 0 && get.sgn(get.value(es[i], ui.selected.targets[0])) == -att2 && get.sgn(get.value(es[i], target)) == sgnatt && target.canEquip(es[i])) {
-						return Math.abs(att);
-					}
-				}
-				if (
-					i == es.length &&
-					(_status.event.nojudge ||
-						!ui.selected.targets[0].countCards("j", function (card) {
-							return target.canAddJudge(card);
-						}))
-				) {
-					return 0;
-				}
-				return -att * get.attitude(player, ui.selected.targets[0]);
-			});
-			next.set("multitarget", true);
-			next.set("targetprompt", _status.event.targetprompt || ["被移走", "移动目标"]);
-			next.set("prompt", event.prompt || "移动场上的一张牌");
-			if (event.prompt2) next.set("prompt2", event.prompt2);
-			if (event.forced) next.set("forced", true);
-			"step 1";
-			event.result = result;
-			if (result.bool) {
-				player.line2(result.targets, "green");
-				event.targets = result.targets;
-			} else {
-				event.finish();
-			}
-			"step 2";
-			game.delay();
-			"step 3";
-			if (targets.length == 2) {
-				player
-					.choosePlayerCard(
-						"ej",
-						true,
-						function (button) {
-							var player = _status.event.player;
-							var targets0 = _status.event.targets0;
-							var targets1 = _status.event.targets1;
-							if (get.attitude(player, targets0) > 0 && get.attitude(player, targets1) < 0) {
-								if (get.position(button.link) == "j") return 12;
-								if (get.value(button.link, targets0) < 0) return 10;
-								return 0;
-							} else {
-								if (get.position(button.link) == "j") return -10;
-								return get.equipValue(button.link);
-							}
-						},
-						targets[0]
-					)
-					.set("nojudge", event.nojudge || false)
-					.set("targets0", targets[0])
-					.set("targets1", targets[1])
-					.set("filterButton", function (button) {
-						var targets1 = _status.event.targets1;
-						if (get.position(button.link) == "j") {
-							if (_status.event.nojudge) return false;
-							return targets1.canAddJudge(button.link);
-						} else {
-							return targets1.canEquip(button.link);
-						}
-					});
-			} else {
-				event.finish();
-			}
-			"step 4";
-			if (result.bool && result.links.length) {
-				player.loseHp();
-			}
-			"step 5";
-			if (result.bool && result.links.length) {
-				var link = result.links[0];
-				if (get.position(link) == "e") {
-					event.targets[1].equip(link);
-				} else if (link.viewAs) {
-					event.targets[1].addJudge({ name: link.viewAs }, [link]);
-				} else {
-					event.targets[1].addJudge(link);
-				}
-				event.targets[0].$give(link, event.targets[1]);
-				event.result.card = link;
-				event.result.position = get.position(link);
-				game.delay();
-				player.addTempSkill("xinzhilve_mark");
-				player.addMark("xinzhilve_mark", 1, false);
-			}
-		},
-	},
-	xinzhilve_use: {
-		audio: "zhilve",
-		filterCard: function () {
-			return false;
-		},
-		selectCard: -1,
-		filterTarget: function (card, player, target) {
-			return player.canUse({ name: "sha", isCard: true }, target, false);
-		},
-		content: function () {
-			player.loseHp();
-			player.draw();
-			player.useCard({ name: "sha", isCard: true }, false, target).forceDie = true;
-			player.addTempSkill("xinzhilve_mark");
-			player.addMark("xinzhilve_mark", 1, false);
-		},
-		ai: {
-			result: {
-				target: function (player, target) {
-					return get.effect(target, { name: "sha" }, player, target);
-				},
-			},
-		},
-	},
-	xinzhilve_mark: {
-		intro: { content: "本回合手牌上限+#" },
-		onremove: true,
-		charlotte: true,
-		mod: {
-			maxHandcard: function (player, num) {
-				return num + player.countMark("xinzhilve_mark");
-			},
-		},
-	},
-	xinxhzhiyan: {
-		audio: "xhzhiyan",
-		enable: "phaseUse",
-		filter: function (event, player) {
-			return (!player.hasSkill("xinxhzhiyan_true") && player.countCards("h") > player.hp) || (!player.hasSkill("xinxhzhiyan_false") && player.countCards("h") < player.maxHp);
-		},
-		filterCard: true,
-		selectCard: function () {
-			var player = _status.event.player;
-			if (player.hasSkill("xinxhzhiyan_true")) return 0;
-			var num = Math.max(0, player.countCards("h") - player.hp);
-			if (ui.selected.cards.length || player.hasSkill("xinxhzhiyan_false") || player.countCards("h") >= player.maxHp) return [num, num];
-			return [0, num];
-		},
-		filterTarget: lib.filter.notMe,
-		selectTarget: function () {
-			if (ui.selected.cards.length) return [1, 1];
-			return [0, 0];
-		},
-		check: function (card) {
-			var player = _status.event.player;
-			var checkx = function (card) {
-				if (
-					player.getUseValue(card, null, true) <= 0 &&
-					game.hasPlayer(function (current) {
-						return current != player && get.value(card, current) > 0 && get.attitude(player, current) > 0;
-					})
-				)
-					return 2;
-				return 1;
-			};
-			if (
-				player.countCards("h", function (card) {
-					return checkx(card) > 0;
-				}) <
-				player.countCards("h") - player.hp
-			)
-				return 0;
-			return checkx(card);
-		},
-		delay: false,
-		discard: false,
-		lose: false,
-		content: function () {
-			var bool = cards && cards.length > 0;
-			player.addTempSkill("xinxhzhiyan_" + bool, "phaseUseEnd");
-			if (!bool) {
-				player.draw(player.maxHp - player.countCards("h"));
-			} else {
-				player.give(cards, target);
-			}
-		},
-		ai: {
-			order: function (obj, player) {
-				if (player.countCards("h") > player.hp) return 10;
-				return 0.5;
-			},
-			result: {
-				player: function (player, target) {
-					if (!ui.selected.cards.length && player.countCards("h") < player.maxHp) return 1;
-					return 0;
-				},
-				target: 1,
-			},
-		},
-	},
-	xinxhzhiyan_true: {},
-	xinxhzhiyan_false: {
-		mod: {
-			playerEnabled: function (card, player, target) {
-				if (player != target && (!get.info(card) || !get.info(card).singleCard || !ui.selected.targets.length)) return false;
-			},
-		},
-		mark: true,
-		intro: {
-			content: "不能对其他角色使用牌",
-		},
-	},
-	weifeng: {
-		audio: 2,
-		trigger: { player: "useCardAfter" },
-		forced: true,
-		filter: function (event, player) {
-			if (
-				!event.targets ||
-				!event.targets.filter(function (target) {
-					return target != player && !target.storage.weifeng2;
-				}).length
-			)
-				return false;
-			var evt = event.getParent("phaseUse");
-			if (!evt || evt.player != player) return false;
-			if (!get.tag(event.card, "damage")) return false;
-			if (!["basic", "trick"].includes(get.type(event.card))) return false;
-			return (
-				player
-					.getHistory("useCard", function (ev) {
-						return ev.getParent("phaseUse") == evt && get.tag(ev.card, "damage") && ["basic", "trick"].includes(get.type(ev.card));
-					})
-					.indexOf(event) == 0 &&
-				game.hasPlayer(function (current) {
-					return current != player && !current.storage.weifeng2 && event.targets.includes(current);
-				})
-			);
-		},
-		content: function () {
-			"step 0";
-			player
-				.chooseTarget(true, "威风：请选择一个目标，令其获得一个【惧(" + get.translation(trigger.card.name) + ")】标记", function (card, player, target) {
-					return player != target && !target.storage.weifeng2 && _status.event.getTrigger().targets.includes(target);
-				})
-				.set("ai", function (target) {
-					return -get.attitude(_status.event.player, target);
-				});
-			"step 1";
-			if (result.bool) {
-				var target = result.targets[0];
-				target.storage.weifeng2 = trigger.card.name;
-				player.line(target, "green");
-				game.log(target, "获得了一个", "#g【惧(" + get.translation(trigger.card.name) + ")】", "标记");
-				target.markSkill("weifeng2");
-				player.addSkill("weifeng3");
-			}
-		},
-	},
-	weifeng2: {
-		intro: {
-			content: "当前“惧”标记名称：$",
-			onunmark: function (storage, player) {
-				if (player.storage.weifeng2) {
-					game.log(player, "移去了一个", "#g【惧(" + get.translation(player.storage.weifeng2) + ")】", "标记");
-					delete player.storage.weifeng2;
-				}
-			},
-		},
-		marktext: "惧",
-	},
-	weifeng3: {
-		trigger: {
-			global: "damageBegin3",
-			player: ["phaseZhunbeiBegin", "dieBegin"],
-		},
-		forced: true,
-		popup: false,
-		filter: function (event, player) {
-			if (event.name != "damage") return true;
-			return event.player != player && typeof event.player.storage.weifeng2 == "string";
-		},
-		content: function () {
-			if (trigger.name == "damage") {
-				player.logSkill("weifeng", trigger.player);
-				if (trigger.card && trigger.card.name == trigger.player.storage.weifeng2) trigger.num++;
-				else if (trigger.player.countGainableCards(player, "he") > 0) player.gainPlayerCard(trigger.player, "he", true);
-				trigger.player.unmarkSkill("weifeng2");
-			} else {
-				game.countPlayer(function (current) {
-					if (current.storage.weifeng2) current.unmarkSkill("weifeng2");
-				});
-				player.removeSkill("weifeng3");
-			}
-		},
-	},
-	gnjinfan: {
-		trigger: { player: "phaseDiscardBegin" },
-		direct: true,
-		locked: false,
-		audio: 2,
-		filter: function (event, player) {
-			var list = [];
-			player.getCards("s", function (card) {
-				if (card.hasGaintag("gnjinfan")) list.add(get.suit(card));
-			});
-			if (list.length >= lib.suit.length) return false;
-			return (
-				player.countCards("h", function (card) {
-					return _status.connectMode || !list.includes(get.suit(card));
-				}) > 0
-			);
-		},
-		content: function () {
-			"step 0";
-			player
-				.chooseCard(
-					"h",
-					get.prompt("gnjinfan"),
-					"将任意张手牌当做“铃”置于武将牌上",
-					[
-						1,
-						(function () {
-							var list = [];
-							var list2 = [];
-							player.getCards("s", function (card) {
-								if (card.hasGaintag("gnjinfan")) list.add(get.suit(card));
-							});
-							player.getCards("h", function (card) {
-								list2.add(get.suit(card));
-							});
-							list2.removeArray(list);
-							return Math.max(1, list2.length);
-						})(),
-					],
-					function (card, player) {
-						return (
-							!player.countCards("s", function (cardx) {
-								return cardx.hasGaintag("gnjinfan") && get.suit(cardx, false) == get.suit(card, player);
-							}) &&
-							!ui.selected.cards.filter(function (cardx) {
-								return get.suit(cardx, player) == get.suit(card, player);
-							}).length
-						);
-					}
-				)
-				.set("ai", function (card) {
-					var player = _status.event.player;
-					if (player.hasUseTarget(card) && !player.hasValueTarget(card)) return 0;
-					if (["sha", "shan", "wuxie", "caochuan"].includes(card.name)) return 2 + Math.random();
-					return 1 + Math.random();
-				})
-				.set("complexCard", true);
-			"step 1";
-			if (result.bool) {
-				player.logSkill("gnjinfan");
-				game.log(player, "将", result.cards, "放到了武将牌上");
-				player.loseToSpecial(result.cards, "gnjinfan").visible = true;
-			} else event.finish();
-			"step 2";
-			player.markSkill("gnjinfan");
-		},
-		group: ["gnjinfan_gain"],
-		marktext: "铃",
-		intro: {
-			mark: function (dialog, storage, player) {
-				dialog.addAuto(
-					player.getCards("s", function (card) {
-						return card.hasGaintag("gnjinfan");
-					})
-				);
-			},
-			markcount: function (storage, player) {
-				return player.getCards("s", function (card) {
-					return card.hasGaintag("gnjinfan");
-				}).length;
-			},
-			onunmark: function (storage, player) {
-				var cards = player.getCards("s", function (card) {
-					return card.hasGaintag("gnjinfan");
-				});
-				if (cards.length) {
-					player.lose(cards, ui.discardPile);
-					player.$throw(cards, 1000);
-					game.log(cards, "进入了弃牌堆");
-				}
-			},
-		},
-		mod: {
-			aiOrder: function (player, card, num) {
-				if (get.itemtype(card) == "card" && card.hasGaintag("gnjinfan")) return num + 0.5;
-			},
-		},
-	},
-	gnjinfan_gain: {
-		audio: "gnjinfan",
-		trigger: { player: "loseAfter" },
-		forced: true,
-		filter: function (event, player) {
-			if (!event.ss || !event.ss.length) return false;
-			for (var i in event.gaintag_map) {
-				if (event.gaintag_map[i].includes("gnjinfan")) return true;
-				return false;
-			}
-		},
-		content: function () {
-			"step 0";
-			var cards = [];
-			for (var i of trigger.ss) {
-				if (!trigger.gaintag_map[i.cardid] || !trigger.gaintag_map[i.cardid].includes("gnjinfan")) continue;
-				var suit = get.suit(i, false);
-				var card = get.cardPile2(function (card) {
-					return !cards.includes(card) && get.suit(card, false) == suit;
-				});
-				if (card) cards.push(card);
-			}
-			if (cards.length) player.gain(cards, "gain2");
-			var num = player.getCards("s", function (card) {
-				return card.hasGaintag("gnjinfan");
-			}).length;
-			if (num) player.markSkill("gnjinfan");
-			else player.unmarkSkill("gnjinfan");
-			"step 1";
-			game.updateRoundNumber();
-		},
-	},
-	gnsheque: {
-		audio: 2,
-		trigger: { global: "phaseZhunbeiBegin" },
-		direct: true,
-		filter: function (event, player) {
-			return event.player.isIn() && event.player.countCards("e") > 0 && lib.filter.targetEnabled({ name: "sha" }, player, event.player) && (player.hasSha() || (_status.connectMode && player.countCards("h") > 0));
-		},
-		content: function () {
-			player
-				.chooseToUse(
-					function (card, player, event) {
-						if (get.name(card) != "sha") return false;
-						return lib.filter.filterCard.apply(this, arguments);
-					},
-					"射却：是否对" + get.translation(trigger.player) + "使用一张杀？"
-				)
-				.set("logSkill", "gnsheque")
-				.set("complexSelect", true)
-				.set("filterTarget", function (card, player, target) {
-					if (target != _status.event.sourcex && !ui.selected.targets.includes(_status.event.sourcex)) return false;
-					return lib.filter.targetEnabled.apply(this, arguments);
-				})
-				.set("sourcex", trigger.player)
-				.set("oncard", function (card) {
-					try {
-						card.gnsheque_tag = true;
-					} catch (e) {
-						alert("发生了一个导致【射却】无法正常触发无视防具效果的错误。请关闭十周年UI/手杀ui等扩展以解决");
-					}
-				});
-		},
-		ai: {
-			unequip: true,
-			unequip_ai: true,
-			skillTagFilter: function (player, tag, arg) {
-				if (tag == "unequip_ai") {
-					if (_status.event.getParent().name != "gnsheque") return false;
-				} else if (!arg || !arg.card || !arg.card.gnsheque_tag) return false;
-			},
 		},
 	},
 	_doublegroup_choice: {
