+++ character/mobile/translate.js
@@ -271,35 +271,6 @@ const translates = {
 	zhengjing_info: "出牌阶段，你可以整理卡牌。然后，你将整理出的卡牌中的至少一张作为“经”置于一名角色的武将牌上，然后获得其余的牌。该角色的准备阶段获得这些牌，且跳过此回合的判定和摸牌阶段。",
 	zhengjing2: "整经",
 
-	mobile_yijiang: "将星独具",
-	yj_zhanghe: "☆张郃",
-	yj_zhanghe_prefix: "☆",
-	yj_zhangliao: "☆张辽",
-	yj_zhangliao_prefix: "☆",
-	yj_xuhuang: "☆徐晃",
-	yj_xuhuang_prefix: "☆",
-	yj_ganning: "☆甘宁",
-	yj_ganning_prefix: "☆",
-	xhzhiyan: "治严",
-	xhzhiyan_info: "出牌阶段，若你的手牌数不等于体力上限，则你可以将手牌摸至/弃至手牌上限，然后本回合不能对其他角色使用牌/可以将弃置的一张牌交给一名其他角色。",
-	zhilve: "知略",
-	zhilve_info: "锁定技，准备阶段，你选择一项：1.移动场上装备区的一张牌并失去1点体力。2.移动场上判定区的一张牌并令本回合手牌上限-1。3.本回合摸牌阶段多摸一张牌且使用的第一张【杀】无距离限制且不计入次数限制。",
-	xinzhilve: "知略",
-	xinzhilve_backup: "知略",
-	xinzhilve_mark: "知略",
-	xinzhilve_info: "出牌阶段限一次，你可以失去1点体力并选择一项：1.移动场上的一张牌；2.视为使用一张无距离限制且不计入次数限制的【杀】并摸一张牌。然后你本回合的手牌上限+1。",
-	xinxhzhiyan: "治严",
-	xinxhzhiyan_false: "治严",
-	xinxhzhiyan_info: "出牌阶段每项各限一次，若你的手牌数：大于体力值，则你可以将X张手牌交给一名其他角色（X为你的手牌数与体力值之差）；小于体力上限，则你可以摸X张牌且本阶段内不能再对其他角色使用牌。（X为你的手牌数与体力上限之差）",
-	weifeng: "威风",
-	weifeng2: "威风",
-	weifeng3: "威风",
-	weifeng_info: "锁定技，当你于出牌阶段内使用第一张伤害性基本牌或普通锦囊牌后，你令此牌的一名没有“惧”的其他目标角色获得一枚名称为此牌牌名的“惧”。有“惧”的角色受到伤害时，其移去“惧”，然后若造成伤害的牌名称和“惧”：相同，此伤害+1；不同，你获得该角色的一张牌。准备阶段开始时或你死亡时，你移去场上的所有“惧”。",
-	gnjinfan: "锦帆",
-	gnjinfan_gain: "锦帆",
-	gnjinfan_info: "弃牌阶段开始时，你可将任意张手牌置于武将牌上，称为“铃”（每种花色的“铃”限一张）。你可以如手牌般使用或打出“铃”。当有“铃”移动到处理区后，你从牌堆中获得与“铃”花色相同的一张牌。",
-	gnsheque: "射却",
-	gnsheque_info: "一名其他角色的准备阶段开始时，若其装备区内有牌，则你可以对其使用一张【杀】（无距离关系的限制且无视防具）。",
 	sp_sufei: "手杀苏飞",
 	sp_sufei_prefix: "手杀",
 	zhengjian: "诤荐",
@@ -531,14 +502,6 @@ const translates = {
 	rezongxuan: "纵玄",
 	rezongxuan_place: "纵玄",
 	rezongxuan_info: "当你的牌因弃置而进入弃牌堆后，你可以将其以任意顺序置于牌堆顶。出牌阶段限一次，你可以摸一张牌，然后将一张牌置于牌堆顶。",
-	yj_huangzhong: "☆黄忠",
-	yj_huangzhong_prefix: "☆",
-	spshidi: "势敌",
-	spshidi_info: "转换技，锁定技。①准备阶段/结束阶段开始时，若你发动此分支的累计次数为奇数/偶数，则你获得一个“☯”。②若你的“☯”数为偶数，则你至其他角色的距离-1，且你使用的黑色【杀】不可被响应。③若你的“☯”数为奇数，则其他角色至你的距离+1，且你不可响应红色【杀】。",
-	spyishi: "义释",
-	spyishi_info: "当你对装备区有牌的其他角色造成伤害时，你可令此伤害-1，然后获得其装备区内的一张牌。",
-	spqishe: "骑射",
-	spqishe_info: "你可以将一张装备牌当做【酒】使用。你的手牌上限+X（X为你装备区内的牌数）。",
 	sp_maojie: "毛玠",
 	bingqing: "秉清",
 	bingqing_info: "当你于出牌阶段内使用的牌结算完毕后，若你于本阶段内此前使用的所有牌的花色与此牌花色均不相同，则你可根据X的值执行对应效果：为2，你令一名角色摸两张牌；为3，你弃置一名角色区域内的一张牌；为4，你对一名其他角色造成1点伤害。（X为你于本阶段内此前使用的所有牌包含的花色数+1）",
@@ -669,12 +632,6 @@ const translates = {
 	xin_caoxiu_prefix: "手杀界",
 	xinqingxi: "倾袭",
 	xinqingxi_info: "每回合限一次。当你对其他角色造成伤害时，你可以令其选择一项：1.弃置X张手牌（X为4减去你至其的距离，至少为1）；2.令此伤害+1。",
-	yj_weiyan: "☆魏延",
-	yj_weiyan_prefix: "☆",
-	mbguli: "孤厉",
-	mbguli_info: "出牌阶段限一次。你可以将所有手牌当做一张无视防具的【杀】使用。此牌结算结束后，若此牌造成过伤害，你可以失去1点体力并将手牌摸至X张（X为你的体力上限）。",
-	mbaosi: "骜肆",
-	mbaosi_info: "锁定技。当你于出牌阶段对一名攻击范围内的角色造成伤害后，你于此阶段对其使用牌无次数限制。",
 	xin_zhoutai: "手杀界周泰",
 	xin_zhoutai_prefix: "手杀界",
 	re_yanwen: "手杀界颜良文丑",
@@ -749,12 +706,6 @@ const translates = {
 	yichong_info: "①准备阶段，你可以选择一名其他角色并选择一个花色，然后你获得其所有此花色的装备牌和其一张此花色的手牌，移除场上的所有“雀”标记，令其获得“雀”标记直到你的下个回合开始。②拥有“雀”标记的角色获得你最后一次发动〖易宠①〗选择的花色的牌后，你获得这些牌（你至多通过每个“雀”得到一张牌）。",
 	wufei: "诬诽",
 	wufei_info: "若场上存在拥有“雀”标记的角色A，则：①当你使用【杀】或伤害类锦囊牌指定第一个目标后，你令A成为此牌伤害来源。②当你受到伤害后，若A的体力值大于3，则你可以令A受到1点无来源伤害。",
-	yj_zhoubuyi: "☆周不疑",
-	yj_zhoubuyi_prefix: "☆",
-	mbhuiyao: "慧夭",
-	mbhuiyao_info: "出牌阶段限一次。你可以受到1点无来源伤害，然后你选择一名其他角色，令其视为对另一名角色造成过1点伤害。",
-	mbquesong: "雀颂",
-	mbquesong_info: "一名角色的结束阶段，若你于本回合受到过伤害，你可以令一名角色选择一项：1.摸X张牌并复原武将牌（X为3，若其装备区非宝物牌牌数不小于三张则X为2）；2.回复1点体力。",
 	xin_yuanshao: "手杀界袁绍",
 	xin_yuanshao_prefix: "手杀界",
 	re_baosanniang: "手杀鲍三娘",
