+++ character/mobile/voices.js
@@ -86,11 +86,6 @@ export default {
 	"#luanqun1": "年过杖朝，自是从心所欲，何来逾矩之理？",
 	"#luanqun2": "位居执慎，博涉多闻，更应秉性而论！",
 	"#laimin:die": "狂嚣之言，一言十过啊……",
-	"#mbhuiyao1": "幸有仓舒为伴，吾不至居高寡寒。",
-	"#mbhuiyao2": "通悟而无笃学之念，则必盈天下之叹也。",
-	"#mbquesong1": "承白雀之瑞，显周公之德。",
-	"#mbquesong2": "挽汉室于危亡，继光武之中兴。",
-	"#yj_zhoubuyi:die": "慧童亡，天下伤……",
 	"#yichong1": "处椒房之尊，得陛下隆宠！",
 	"#yichong2": "三千宠爱？当聚于我一身！",
 	"#wufei1": "巫蛊实乃凶邪之术，陛下不可不察！",
@@ -123,11 +118,6 @@ export default {
 	"#chengzhang1": "弦急悲声发，聆我慷慨言。",
 	"#chengzhang2": "盛时不再来，百年忽我遒。",
 	"#re_caozhi:die": "先民谁不死，知命复何忧？",
-	"#mbguli1": "今若弑此昏聩主，纵蒙恶名又如何？",
-	"#mbguli2": "韩玄少谋多忌，吾今当诛之！",
-	"#mbaosi1": "凶慢骜肆，天生狂骨！",
-	"#mbaosi2": "暴戾恣睢，傲视诸雄！",
-	"#yj_weiyan:die": "使君为何弃我而去……呃啊！",
 	"#redingpin1": "察举旧制已隳，简拔当立中正。",
 	"#redingpin2": "置州郡中正，以九品进退人才。",
 	"#refaen1": "法不可容之事，情或能原宥。",
@@ -242,13 +232,6 @@ export default {
 	"#yingfeng1": "公内奉天子，外令诸侯，则天下可定。",
 	"#yingfeng2": "公怀奉君之心，则民有安固之志。",
 	"#sp_maojie:die": "废立大事，公不可不慎……",
-	"#spshidi1": "诈败以射之，其必死矣！",
-	"#spshidi2": "呃啊，中其拖刀计矣！",
-	"#spyishi1": "昨日释忠之恩，今吾虚射以报。",
-	"#spyishi2": "君刀不砍头颅，吾箭只射盔缨。",
-	"#spqishe1": "诱敌之计已成，吾且拈弓搭箭！",
-	"#spqishe2": "关羽即至吊桥，既已控弦，如何是好？",
-	"#yj_huangzhong:die": "关云长义释黄某，吾又安忍射之……",
 	"#zhiyan_re_yufan1": "失忠于信，何以侍君！",
 	"#zhiyan_re_yufan2": "恕臣耿直，犯言谏正！",
 	"#rezongxuan1": "凡夫俗子，怎识得卦象玄妙？",
@@ -418,20 +401,6 @@ export default {
 	"#gaoyuan1": "还请告知兴霸，请他务必相助。",
 	"#gaoyuan2": "如今事急，唯有兴霸可救。",
 	"#sp_sufei:die": "本可共图大业，奈何主公量狭器小啊……",
-	"#weifeng1": "广散惧义，尽泄敌之斗志。",
-	"#weifeng2": "哼哼，若尔等惧我，自当卷甲以降！",
-	"#yj_zhangliao:die": "惑于女子而尽失战机，庸主误我啊……",
-	"#zhilve1": "将者，上不治天，下不治地，中不治人。",
-	"#zhilve2": "料敌之计，明敌之意，因况反制。",
-	"#yj_zhanghe:die": "若非小人作梗，何至官渡之败……",
-	"#xhzhiyan1": "治军严谨，方得精锐之师。",
-	"#xhzhiyan2": "精兵当严于律己，束身自修。",
-	"#yj_xuhuang:die": "唉，明主未遇，大功未成……",
-	"#gnjinfan1": "扬锦帆，劫四方，快意逍遥！",
-	"#gnjinfan2": "铃声所至之处，再无安宁！",
-	"#gnsheque1": "看我此箭，取那轻舟冒进之人性命！",
-	"#gnsheque2": "纵有劲甲良盾，也难挡我神射之威！",
-	"#yj_ganning:die": "铜铃声……怕是听不到了……",
 	"#yinghun_re_sunjian1": "义定四野，武匡海内。",
 	"#yinghun_re_sunjian2": "江东男儿，皆胸怀匡扶天下之志。",
 	"#re_sunjian:die": "吾身虽死，忠勇须传……",
