+++ character/newjiang/character.js
@@ -1,0 +1,31 @@
+const characters = {
+	yj_sb_guojia: ["male", "wei", 3, ["xianmou", "lunshi"]],
+	xunyuxunyou: ["male", "wei", 3, ["zhinang", "gouzhu"], ["name:荀|彧-荀|攸"]],
+	xiahoumao: ["male", "wei", 4, ["tongwei", "cuguo"], ["name:夏侯|楙"]],
+	chenshi: ["male", "shu", 4, ["qingbei"]],
+	sunli: ["male", "wei", 4, ["kangli"]],
+	feiyao: ["male", "wei", 4, ["zhenfeng"]],
+	wuanguo: ["male", "qun", 4, ["diezhang", "duanwan"]],
+	hanlong: ["male", "wei", 4, ["duwang", "cibei"]],
+	yj_qiaozhou: ["male", "shu", 3, ["shiming", "jiangxi"]],
+	yj_sufei: ["male", "wu", 4, ["shuojian"]],
+	liwan: ["female", "wei", 3, ["liandui", "biejun"]],
+	zhugeshang: ["male", "shu", 3, ["sangu", "yizu"], ["name:诸葛|尚"]],
+	kebineng: ["male", "qun", 4, ["kousheng"]],
+	lukai: ["male", "wu", 4, ["lkbushi", "lkzhongzhuang"]],
+	linghuyu: ["male", "wei", 4, ["xvzhi"], ["name:令狐|愚"]],
+	yj_simafu: ["male", "wei", 3, ["beiyu", "duchi"], ["name:司马|孚"]],
+	yj_xuangongzhu: ["female", "wei", 3, ["yjqimei", "yjzhuiji"], ["name:司马|null"]],
+	xukun: ["male", "wu", 4, ["fazhu"]],
+	yj_zhangliao: ["male", "qun", 4, ["weifeng"]],
+	yj_zhanghe: ["male", "qun", 4, ["xinzhilve"]],
+	yj_xuhuang: ["male", "qun", 4, ["xinxhzhiyan"]],
+	yj_ganning: ["male", "qun", 4, ["gnjinfan", "gnsheque"]],
+	yj_huangzhong: ["male", "qun", 4, ["spshidi", "spyishi", "spqishe"]],
+	yj_weiyan: ["male", "qun", "4/4/1", ["mbguli", "mbaosi"]],
+	yj_zhoubuyi: ["male", "wei", 3, ["mbhuiyao", "mbquesong"]],
+	mp_wangrong: ["male", "wei", 3, ["mpjianlin", "mpsixiao"]],
+	mp_liuling: ["male", "jin", 3, ["mpjiusong", "mpmaotao", "mpbishi"], ["doublegroup:wei:qun:jin"]],
+};
+
+export default characters;
