+++ character/newjiang/characterReplace.js
@@ -1,0 +1,8 @@
+const characterReplaces = {
+	liwan: ["ol_liwan", "liwan"],
+	wuxian: ["wuxian", "clan_wuxian"],
+	simafu: ["mb_simafu", "simafu", "yj_simafu"],
+	xuangongzhu: ["yj_xuangongzhu", "xuangongzhu"],
+};
+
+export default characterReplaces;
