+++ character/newjiang/perfectPairs.js
@@ -1,0 +1,1 @@
+export default {};
