+++ character/newjiang/pinyin.js
@@ -1,0 +1,2 @@
+const pinyins = {};
+export default pinyins;
