+++ character/newjiang/sort.js
@@ -1,0 +1,15 @@
+const characterSort = {
+	yijiang_2020: ["yj_zhanghe", "yj_zhangliao", "yj_xuhuang", "yj_ganning"],
+	yijiang_2022: ["yj_huangzhong", "lukai", "kebineng", "zhugeshang", "liwan", "wuanguo", "hanlong", "yj_sufei", "yj_qiaozhou"],
+	yijiang_2023: ["mp_liuling", "yj_weiyan", "xiahoumao", "chenshi", "sunli", "feiyao"],
+	yijiang_2024: ["mp_wangrong", "yj_zhoubuyi", "linghuyu", "yj_simafu", "yj_xuangongzhu", "xukun", "xunyuxunyou", "yj_sb_guojia"],
+};
+
+const characterSortTranslate = {
+	yijiang_2020: "一将成名2020",
+	yijiang_2022: "一将成名2022",
+	yijiang_2023: "一将成名2023",
+	yijiang_2024: "一将成名2024",
+};
+
+export { characterSort, characterSortTranslate };
