+++ character/newjiang/voices.js
@@ -1,0 +1,88 @@
+export default {
+	"#zhenfeng1": "河西诸贼作乱，吾当驱万里之远。",
+	"#zhenfeng2": "可折诸葛之锋而御者，独我其谁。",
+	"#feiyao:die": "姜维，你果然是蜀军内应！",
+	// 击、逆、泰、每，懂？
+	"#fazhu1": "击风雨于共济，逆流亦溯千帆。",
+	"#fazhu2": "泰山轻于大义，每思志士、何惧临渊。",
+	"#fazhu3": "大江潮起，伐苇成舟，载江南春风。",
+	"#xukun:die": "何处射来的流矢？",
+	"#kangli1": "地界纷争皋陶难断，然图藏天府，坐上可明。",
+	"#kangli2": "正至歉岁，难征百姓于役，望陛下明鉴。",
+	"#sunli:die": "国无矩不立，何谓之方圆？",
+	"#tongwei1": "集关西诸路大军，必雪当年长坂坡之耻。",
+	"#tongwei2": "手织天网十万尺，欲擒飞龙落彀中。",
+	"#cuguo1": "身担父命，怎可蜷于宫阙。",
+	"#cuguo2": "体承国运，岂能缩居朝堂。",
+	"#xiahoumao:die": "一将无能，徒累死三军……",
+	"#qingbei1": "待追上那司马懿，定教他没好果子吃！",
+	"#qingbei2": "身若不周，吾一人可作擎北之柱。",
+	"#chenshi:die": "丞相、丞相！是魏延指使我的！",
+	"#diezhang1": "某家这大锤，舞起来那叫一个万夫莫敌。",
+	"#diezhang2": "贼吕布何在？某家来取汝性命了！",
+	"#duanwan1": "好你个吕奉先，竟敢卸我膀子！",
+	"#duanwan2": "汝这匹夫，为何往手腕上招呼？",
+	"#wuanguo:die": "吕奉先，你给某家等着！",
+	"#duwang1": "此去，欲诛敌莽、杀单于。",
+	"#duwang2": "风萧萧兮易水寒，壮士一去兮不复还！",
+	"#cibei1": "匹夫一怒，流血二人，天下缟素。",
+	"#cibei2": "我欲效专诸、聂政之旧事，逐天狼于西北。",
+	"#hanlong:die": "杀轲比能者，韩龙也！",
+	"#shiming1": "今天命在北，我等已尽人事。",
+	"#shiming2": "益州国疲民敝，非人力可续之。",
+	"#jiangxi1": "典午忽兮，月酉没兮。",
+	"#jiangxi2": "周慕孔子遗风，可与刘、扬同轨。",
+	"#yj_qiaozhou:die": "炎汉百年之业，吾一言毁之……",
+	"#shuojian1": "我数荐卿而祖不用，其之失也。",
+	"#shuojian2": "兴霸乃当世豪杰，何患无爵。",
+	"#yj_sufei:die": "兴霸何在？吾命休矣……",
+	"#liandui1": "以句相联，抒离散之苦。",
+	"#liandui2": "以诗相对，颂哀怨之情。",
+	"#biejun1": "彼岸荼蘼远，落寞北风凉。",
+	"#biejun2": "此去经年，不知何时能归？",
+	"#liwan:die": "生不能同寝，死亦难同穴……",
+	"#sangu1": "思报君恩，尽父子之忠。",
+	"#sangu2": "欲酬三顾，竭三代之力。",
+	"#yizu1": "仿祖父行事，可阻敌袭。",
+	"#yizu2": "习先人故智，可御寇侵。",
+	"#zhugeshang:die": "父子荷国重恩，当尽忠以报！",
+	"#kousheng1": "驰马掠野，塞外称雄。",
+	"#kousheng2": "控弦十万，纵横漠南。",
+	"#kebineng:die": "草原雄鹰，折翼于此……",
+	"#lkbushi1": "论演玄意，以筮辄验。",
+	"#lkbushi2": "手不释书，好研经卷。",
+	"#lkzhongzhuang1": "秽尘天听，卿有不测之祸！",
+	"#lkzhongzhuang2": "倾乱国政，安得寿终正寝？",
+	"#lukai:die": "不听忠言，国将亡矣……",
+	"#mbhuiyao1": "幸有仓舒为伴，吾不至居高寡寒。",
+	"#mbhuiyao2": "通悟而无笃学之念，则必盈天下之叹也。",
+	"#mbquesong1": "承白雀之瑞，显周公之德。",
+	"#mbquesong2": "挽汉室于危亡，继光武之中兴。",
+	"#yj_zhoubuyi:die": "慧童亡，天下伤……",
+	"#mbguli1": "今若弑此昏聩主，纵蒙恶名又如何？",
+	"#mbguli2": "韩玄少谋多忌，吾今当诛之！",
+	"#mbaosi1": "凶慢骜肆，天生狂骨！",
+	"#mbaosi2": "暴戾恣睢，傲视诸雄！",
+	"#yj_weiyan:die": "使君为何弃我而去……呃啊！",
+	"#spshidi1": "诈败以射之，其必死矣！",
+	"#spshidi2": "呃啊，中其拖刀计矣！",
+	"#spyishi1": "昨日释忠之恩，今吾虚射以报。",
+	"#spyishi2": "君刀不砍头颅，吾箭只射盔缨。",
+	"#spqishe1": "诱敌之计已成，吾且拈弓搭箭！",
+	"#spqishe2": "关羽即至吊桥，既已控弦，如何是好？",
+	"#yj_huangzhong:die": "关云长义释黄某，吾又安忍射之……",
+	"#weifeng1": "广散惧义，尽泄敌之斗志。",
+	"#weifeng2": "哼哼，若尔等惧我，自当卷甲以降！",
+	"#yj_zhangliao:die": "惑于女子而尽失战机，庸主误我啊……",
+	"#zhilve1": "将者，上不治天，下不治地，中不治人。",
+	"#zhilve2": "料敌之计，明敌之意，因况反制。",
+	"#yj_zhanghe:die": "若非小人作梗，何至官渡之败……",
+	"#xhzhiyan1": "治军严谨，方得精锐之师。",
+	"#xhzhiyan2": "精兵当严于律己，束身自修。",
+	"#yj_xuhuang:die": "唉，明主未遇，大功未成……",
+	"#gnjinfan1": "扬锦帆，劫四方，快意逍遥！",
+	"#gnjinfan2": "铃声所至之处，再无安宁！",
+	"#gnsheque1": "看我此箭，取那轻舟冒进之人性命！",
+	"#gnsheque2": "纵有劲甲良盾，也难挡我神射之威！",
+	"#yj_ganning:die": "铜铃声……怕是听不到了……",
+};
