+++ character/onlyOL/sort.js
@@ -3,7 +3,11 @@ const characterSort = {
 	onlyOL_yijiang2: ["ol_caozhang", "ol_chengpu", "ol_wangyi", "ol_liubiao"],
 	onlyOL_yijiang3: ["ol_yufan", "ol_liru"],
 	onlyOL_yijiang4: ["ol_caifuren"],
-	onlyOL_sb: ["ol_sb_jiangwei", "ol_sb_guanyu", "ol_sb_taishici", "ol_sb_yuanshao", "ol_sb_pangtong", "ol_sb_kongrong"],
+	onlyOL_sb_mouding: ["ol_sb_jiangwei", "ol_sb_pangtong"],
+	onlyOL_sb_wudong: ["ol_sb_guanyu"],
+	onlyOL_sb_fenwu: ["ol_sb_taishici", "ol_sb_yuanshao"],
+	onlyOL_sb_shiren: ["ol_sb_kongrong"],
+	onlyOL_waitingforsort: [],
 };
 
 const characterSortTranslate = {
@@ -11,7 +15,11 @@ const characterSortTranslate = {
 	onlyOL_yijiang2: "OL专属·将2",
 	onlyOL_yijiang3: "OL专属·将3",
 	onlyOL_yijiang4: "OL专属·将4",
-	onlyOL_sb: "OL专属·上兵伐谋",
+	onlyOL_sb_mouding: "OL专属·谋定天下",
+	onlyOL_sb_wudong: "OL专属·武动乾坤",
+	onlyOL_sb_fenwu: "上兵伐谋·奋武扬威",
+	onlyOL_sb_shiren: "上兵伐谋·施仁布泽",
+	onlyOL_waitingforsort: "等待分包",
 };
 
 export { characterSort, characterSortTranslate };
