+++ character/onlyOL/translate.js
@@ -111,7 +111,7 @@ const translates = {
 	ol_sb_kongrong_prefix: "OL谋",
 	olsbliwen: "立文",
 	olsbliwen_info: "①当你于回合内使用牌时，若此牌与你本回合使用的上一张牌的花色或点数相同，则你获得1枚“贤”标记（一名角色的“贤”标记上限为3）。②回合结束时，你可以失去任意枚“贤”标记并令等量其他角色各获得1枚“贤”标记，然后场上所有有“贤”标记的角色按照“贤”标记由大到小的顺序（“贤”标记数相同的角色按座次逆时针排序）依次选择是否使用一张非装备牌，若不选择使用，则其移去所有“贤”标记，然后你摸等量的牌。",
-	olsbzhengyi: "争议",
+	olsbzhengyi: "争义",
 	olsbzhengyi_info: "当一名有“贤”标记的角色受到伤害时，所有有“贤”标记同时选择是否取消此伤害，所有角色选择完毕后，若有角色选择是，则取消此伤害，然后选择是的角色中体力值最大的角色依次失去X点体力（X为伤害值）。",
 };
