+++ character/xianding/sort.js
@@ -17,7 +17,7 @@ const characterSort = {
 	sp2_zhonghu: ["dc_jiangji", "dc_wangling", "dc_simashi", "dc_caoshuang"],
 	sp2_zijing: ["dc_sb_zhugejin", "dc_sb_guanping"],
 	sp2_dushi: ["dc_sb_caoang", "dc_sb_zhangxiu", "dc_sb_dianwei"],
-	sp2_waitingforsort: [],
+	sp2_waitingforsort: ["liutan"],
 };
 
 const characterSortTranslate = {
