+++ character/xianding/translate.js
@@ -647,10 +647,6 @@ const translates = {
 	dcbizu_info: "出牌阶段，你可以令所有手牌数与你相同的角色摸一张牌。然后若这些角色与本回合此前某次发动〖庇族〗的目标角色均相同，本回合〖庇族〗失效。",
 	dcwuxie: "无胁",
 	dcwuxie_info: "每回合限一次，当你获得牌后，你可以将其中任意张伤害牌置于牌堆底并令一名其他角色弃置等量张牌。",
-	dcxianmou: "先谋",
-	dcxianmou_info: "转换技，你失去过牌的回合结束时，你可以：阴，观看牌堆顶五张牌并获得至多X张牌，若未获得X张牌则获得〖遗计〗直到再发动此项；阳，观看一名角色手牌并弃置其中至多X张牌，若弃置X张牌则你进行一次【闪电】判定。（X为你本回合失去牌数）",
-	dclunshi: "论势",
-	dclunshi_info: "一名角色对其以外的角色使用普通锦囊牌的结算中，若你手牌中两张颜色的牌数量相同，你可将一张手牌当作不可被响应的【无懈可击】使用。",
 	liutan: "柳婒",
 	dcjingyin: "经音",
 	dcjingyin_info: "一名角色于回合外使用【杀】结算完毕后，你可以令一名使用者外的角色获得此【杀】对应的所有实体牌，且其使用这些牌无次数限制。",
