+++ character/yijiang/skill.js
@@ -2,1947 +2,6 @@ import { lib, game, ui, get, ai, _status } from "../../noname.js";
 
 /** @type { importCharacterConfig['skill'] } */
 const skills = {
-	//一将2024
-	//令狐愚
-	xvzhi: {
-		audio: 2,
-		enable: "phaseUse",
-		filter(event, player) {
-			return (
-				game.countPlayer(target => {
-					return lib.skill.xvzhi.filterTarget(null, player, target);
-				}) > 1
-			);
-		},
-		filterTarget(card, player, target) {
-			const stat = player.getStat("xvzhi");
-			return target.countCards("h") && (!stat || !stat.includes(target));
-		},
-		selectTarget: 2,
-		usable: 1,
-		multiline: true,
-		multitarget: true,
-		async content(event, trigger, player) {
-			const targets = event.targets;
-			if (!player.getStat().xvzhi) player.getStat().xvzhi = [];
-			player.getStat().xvzhi.addArray(targets);
-			if (targets.some(i => !i.countCards("h"))) return;
-			const result = await player
-				.chooseCardOL(targets, "h", true, [1, Infinity], "蓄志：选择任意张手牌并与对方交换")
-				.set("ai", card => {
-					const player = get.event("player"),
-						target = get
-							.event()
-							.getParent(2)
-							.targets.find(i => i != player);
-					const sha = new lib.element.VCard({ name: "sha" });
-					const playerEffect = player.hasUseTarget(sha, false)
-						? Math.max(
-								...game
-									.filterPlayer(current => player.canUse(sha, current, false))
-									.map(current => {
-										return get.effect(current, sha, player, player);
-									})
-							)
-						: 0;
-					const targetEffect = target.hasUseTarget(sha, false)
-						? Math.max(
-								...game
-									.filterPlayer(current => target.canUse(sha, current, false))
-									.map(current => {
-										return get.effect(current, sha, player, player);
-									})
-							)
-						: 0;
-					return 5 + 2 * get.sgn(playerEffect - targetEffect) - get.value(card);
-				})
-				.forResult();
-			await targets[0].swapHandcards(targets[1], result[0].cards, result[1].cards);
-			if (result[0].cards.length == result[1].cards.length) {
-				await player.draw(2);
-				player.getStat("skill").xvzhi--;
-			} else {
-				const aim = targets[result[0].cards.length > result[1].cards.length ? 0 : 1];
-				const sha = new lib.element.VCard({ name: "sha" });
-				if (aim.hasUseTarget(sha, false)) {
-					await aim.chooseUseTarget(sha, true, false, "nodistance");
-				}
-			}
-		},
-		ai: {
-			order: 5,
-			result: {
-				target(player, target) {
-					return get.sgn(get.attitude(player, target)) * target.countCards("h");
-				},
-			},
-		},
-	},
-	//司马孚
-	beiyu: {
-		audio: 2,
-		enable: "phaseUse",
-		filter(event, player) {
-			return player.countCards("h") < player.maxHp;
-		},
-		usable: 1,
-		async content(event, trigger, player) {
-			await player.drawTo(player.maxHp);
-			if (!player.countCards("h")) return;
-			const suits = player
-				.getCards("h")
-				.reduce((list, card) => list.add(get.suit(card)), [])
-				.sort((a, b) => lib.suit.indexOf(b) - lib.suit.indexOf(a));
-			const result = await player
-				.chooseControl(suits)
-				.set("prompt", "备预：将一种花色的手牌牌置于牌堆底")
-				.set("ai", () => {
-					const player = get.event("player");
-					let suits = get.event("controls").slice();
-					suits.sort((a, b) => player.countCards("h", { suit: a }) - player.countCards("h", { suit: b }));
-					return suits[0];
-				})
-				.forResult();
-			if (result.control) {
-				const suit = result.control,
-					cards = player.getCards("h", { suit: suit });
-				if (cards.length) {
-					let resultx;
-					if (cards.length == 1) {
-						resultx = { bool: true, moved: [cards] };
-					} else {
-						resultx = await player
-							.chooseToMove("备预：将牌按顺序置于牌堆底", true)
-							.set("list", [["牌堆底", cards]])
-							.set("processAI", list => {
-								return [list[0][1].slice(0)];
-							})
-							.forResult();
-					}
-					if (resultx.bool) {
-						const moved = resultx.moved[0];
-						if (moved.length) {
-							await player.lose(cards, ui.cardPile);
-							for (let i = 0; i < moved.length; i++) {
-								const card = moved[i];
-								card.fix();
-								ui.cardPile.appendChild(card);
-							}
-						}
-					}
-				}
-			}
-		},
-		ai: {
-			order: 0.001,
-			result: { player: 1 },
-		},
-	},
-	duchi: {
-		audio: 2,
-		trigger: { target: "useCardToTargeted" },
-		filter(event, player) {
-			return event.player != player;
-		},
-		usable: 1,
-		logTarget: "player",
-		check(event, player) {
-			return get.effect(player, event.card, event.player, player) <= 0;
-		},
-		async content(event, trigger, player) {
-			await player.draw().set("bottom", true);
-			if (player.countCards("h")) {
-				await player.showHandcards(get.translation(player) + "发动了【督持】");
-				const colors = player.getCards("h").reduce((list, card) => list.add(get.color(card)), []);
-				if (colors.length == 1) {
-					player.popup("洗具");
-					trigger.getParent().excluded.add(player);
-					return;
-				}
-			}
-			player.popup("杯具");
-		},
-		ai: { threaten: 0.8 },
-	},
-	//宣公主
-	yjqimei: {
-		audio: 2,
-		enable: "phaseUse",
-		filter(event, player) {
-			const count = player.getStat("skill").yjqimei;
-			if (count && count > 0 && !player.hasSkill("yjqimei_rewrite")) return false;
-			return true;
-		},
-		filterTarget: lib.filter.notMe,
-		usable: 2,
-		async content(event, trigger, player) {
-			const target = event.target;
-			await player.draw(2, "nodelay");
-			await target.draw(2);
-			const targets = [player, target].filter(current => current.countDiscardableCards(current, "he"));
-			if (targets.length) {
-				const result = await player
-					.chooseCardOL(targets, "he", true, 2, "齐眉：请弃置两张牌", (card, player, target) => {
-						return lib.filter.cardDiscardable(card, player);
-					})
-					.forResult();
-				if (result.length == 1) targets[0].discard(result[0].cards);
-				else {
-					await game
-						.loseAsync({
-							lose_list: [
-								[targets[0], result[0].cards],
-								[targets[1], result[1].cards],
-							],
-						})
-						.setContent("discardMultiple");
-					await game.delayx();
-				}
-				let cards = result.reduce((list, evt) => {
-					list.addArray(evt.cards);
-					return list;
-				}, []);
-				const suits = cards.reduce((list, card) => list.add(get.suit(card)), []);
-				switch (suits.length) {
-					case 1:
-						while (cards.length) {
-							const card = cards.shift();
-							if (player.hasUseTarget(card)) {
-								player.$gain2(card, false);
-								await game.delayx();
-								await player.chooseUseTarget(true, card, false);
-							}
-						}
-						break;
-					case 2:
-						for (const current of [player, target]) {
-							if (!current.isIn()) continue;
-							if (current.isLinked()) await current.link(false);
-							if (current.isTurnedOver()) await current.turnOver(false);
-						}
-						break;
-					case 3:
-						for (const current of [player, target]) {
-							if (current.isIn() && !current.isLinked()) await current.link(true);
-						}
-						break;
-					case 4:
-						await player.draw("nodelay");
-						await target.draw();
-						player.addTempSkill("yjqimei_rewrite");
-						break;
-				}
-			}
-		},
-		ai: {
-			order: 9,
-			result: {
-				target(player, target) {
-					const att = get.sgn(get.attitude(player, target));
-					return (2 + att) * att;
-				},
-			},
-		},
-		subSkill: { rewrite: { charlotte: true } },
-	},
-	yjzhuiji: {
-		audio: 2,
-		trigger: { player: "die" },
-		filter(event, player) {
-			return game.hasPlayer(target => {
-				return (
-					target != player &&
-					Array.from({ length: 5 })
-						.map((_, i) => i + 1)
-						.some(i => target.hasEmptySlot(i))
-				);
-			});
-		},
-		forceDie: true,
-		skillAnimation: true,
-		animationColor: "water",
-		async cost(event, trigger, player) {
-			event.result = await player
-				.chooseTarget(get.prompt2("yjzhuiji"), (card, player, target) => {
-					return (
-						target != player &&
-						Array.from({ length: 5 })
-							.map((_, i) => i + 1)
-							.some(i => target.hasEmptySlot(i))
-					);
-				})
-				.set("ai", target => {
-					const player = get.event("player");
-					return (
-						get.sgn(get.attitude(player, target)) *
-						Array.from({ length: 5 })
-							.map((_, i) => i + 1)
-							.reduce((sum, i) => sum + target.countEmptySlot(i), 0)
-					);
-				})
-				.set("forceDie", true)
-				.forResult();
-		},
-		async content(event, trigger, player) {
-			const target = event.targets[0];
-			let num = 1,
-				cards = [];
-			while (num <= 5) {
-				while (target.hasEmptySlot(num)) {
-					const card = get.cardPile2(card => {
-						return !cards.includes(card) && get.subtype(card) == "equip" + num && target.canUse(card, target);
-					});
-					if (card) {
-						cards.push(card);
-						target.$gain2(card, false);
-						await game.delayx();
-						await target.chooseUseTarget(card, true, "nopopup");
-					} else break;
-				}
-				num++;
-			}
-			if (cards.length) {
-				target.addSkill("yjzhuiji_buff");
-				target.markAuto("yjzhuiji_buff", cards);
-			}
-		},
-		subSkill: {
-			buff: {
-				charlotte: true,
-				mod: {
-					aiValue(player, card, num) {
-						if (player.getStorage("yjzhuiji_buff").includes(card)) return num + 100;
-					},
-					aiUseful(player, card, num) {
-						if (player.getStorage("yjzhuiji_buff").includes(card)) return num / 114514;
-					},
-				},
-				trigger: {
-					player: "loseAfter",
-					global: ["equipAfter", "addJudgeAfter", "gainAfter", "loseAsyncAfter", "addToExpansionAfter"],
-				},
-				filter(event, player) {
-					const evt = event.getl(player);
-					return evt && evt.es && evt.es.some(i => player.getStorage("yjzhuiji_buff").includes(i));
-				},
-				forced: true,
-				popup: false,
-				firstDo: true,
-				content() {
-					const evt = trigger.getl(player);
-					const cards = evt.es.filter(i => player.getStorage("yjzhuiji_buff").includes(i));
-					player.unmarkAuto("yjzhuiji_buff", cards);
-					for (const card of cards) player.disableEquip(get.subtype(card));
-				},
-				intro: {
-					mark(dialog, storage) {
-						if (storage && storage.length) dialog.addSmall([storage, "vcard"]);
-						else return "暂无装备";
-					},
-				},
-			},
-		},
-	},
-	//徐琨（菜不菜我不知道）
-	fazhu: {
-		audio: 3,
-		trigger: { player: "phaseZhunbeiBegin" },
-		filter(event, player) {
-			return player.hasCard(card => !get.tag(card, "damage") && player.canRecast(card), "hej");
-		},
-		async cost(event, trigger, player) {
-			event.result = await player
-				.choosePlayerCard(get.prompt(event.name.slice(0, -5)), player, "hej", [1, Infinity])
-				.set("ai", button => {
-					const card = button.link;
-					if (get.position(card) == "j") return 10;
-					return 6 - get.value(card);
-				})
-				.set("filterButton", button => {
-					const card = button.link;
-					return !get.tag(card, "damage") && player.canRecast(card);
-				})
-				.forResult();
-		},
-		async content(event, trigger, player) {
-			await player.recast(event.cards);
-			const cards = player
-				.getHistory("gain", evt => evt.getParent(3) == event)
-				.reduce((list, evt) => {
-					list.addArray(evt.cards);
-					return list;
-				}, []);
-			let num = Math.min(cards.length, game.countPlayer()),
-				list = [];
-			if (!num) return;
-			if (_status.connectMode) game.broadcastAll(() => (_status.noclearcountdown = true));
-			while (num - list.length > 0 && cards.some(i => get.owner(i) == player && get.position(i) == "h" && !i.hasGaintag("olsujian_given"))) {
-				const result = await player
-					.chooseCardTarget({
-						prompt: "筏铸：将以此法获得的牌交给任意角色各一张",
-						position: "he",
-						animate: false,
-						filterCard(card, player) {
-							if (!get.event("cards").includes(card)) return false;
-							return !get.event("list").some(list => list[1] == card);
-						},
-						filterTarget(card, player, target) {
-							return !get.event("list").some(list => list[0] == target);
-						},
-						ai1(card) {
-							if (card.name == "sha") return 2.5;
-							return 1 + Math.random();
-						},
-						ai2(target) {
-							return get.attitude(get.event("player"), target);
-						},
-					})
-					.set("forced", !list.length)
-					.set("list", list)
-					.set("cards", cards)
-					.forResult();
-				if (result.bool) {
-					list.push([result.targets[0], result.cards[0]]);
-					player.addGaintag(result.cards, "olsujian_given");
-				} else break;
-			}
-			if (_status.connectMode) {
-				game.broadcastAll(() => {
-					delete _status.noclearcountdown;
-					game.stopCountChoose();
-				});
-			}
-			if (list.length) {
-				const targets = list.slice().map(list => list[0]);
-				await game
-					.loseAsync({
-						gain_list: list,
-						player: player,
-						cards: list.slice().map(list => list[1]),
-						giver: player,
-						animate: "giveAuto",
-					})
-					.setContent("gaincardMultiple");
-				for (const target of targets) {
-					await target
-						.chooseToUse(function (card, player, event) {
-							if (get.name(card) != "sha") return false;
-							return lib.filter.cardEnabled.apply(this, arguments);
-						})
-						.set("openskilldialog", "筏铸：是否使用一张【杀】（无距离限制）？")
-						.set("norestore", true)
-						.set("custom", {
-							add: {},
-							replace: { window: function () {} },
-						})
-						.set("targetRequired", true)
-						.set("complexSelect", true)
-						.set("filterTarget", function (card, player, target) {
-							return lib.filter.targetEnabled.apply(this, arguments);
-						})
-						.set("addCount", false);
-				}
-			}
-		},
-		ai: {
-			effect: {
-				target(card, player, target) {
-					if (!card || get.type(card) != "delay") return;
-					if (!get.tag(card, "damage") && target.canRecast(card)) return "zeroplayertarget";
-				},
-			},
-		},
-	},
-	//一将2023
-	//孙礼
-	kangli: {
-		audio: 2,
-		trigger: {
-			player: "damageEnd",
-			source: "damageSource",
-		},
-		forced: true,
-		locked: false,
-		content: function () {
-			player.draw(2).gaintag = ["kangli"];
-			player.when({ source: "damageBegin1" }).then(() => {
-				var cards = player.getCards("h", card => card.hasGaintag("kangli") && lib.filter.cardDiscardable(card, player, "kangli"));
-				if (cards.length) player.discard(cards);
-			});
-		},
-		ai: {
-			maixie: true,
-		},
-	},
-	//夏侯楙
-	tongwei: {
-		audio: 2,
-		enable: "phaseUse",
-		usable: 1,
-		filter: function (event, player) {
-			return player.countCards("he", card => lib.skill.tongwei.filterCard(card, player)) > 1 && game.hasPlayer(i => i != player);
-		},
-		filterTarget: lib.filter.notMe,
-		filterCard: lib.filter.cardRecastable,
-		selectCard: 2,
-		position: "he",
-		discard: false,
-		lose: false,
-		delay: false,
-		popname: true,
-		check: function (card) {
-			var num = 6.5;
-			if (ui.selected.cards.length) {
-				var cardx = ui.selected.cards[0];
-				num = get.number(cardx);
-			}
-			var del = Math.abs(get.number(card) - num);
-			return 5 + del / 5 - get.value(card);
-		},
-		content: function () {
-			"step 0";
-			player.recast(cards);
-			"step 1";
-			var numbers = cards.map(c => get.number(c, player)).sort((a, b) => a - b);
-			target.when("useCard1").then(() => {
-				trigger._tongwei_checked = true;
-			});
-			target
-				.when("useCardAfter")
-				.assign({
-					numbers: numbers,
-					playerx: player,
-					mod: {
-						aiOrder: function (player, card, num) {
-							var number = get.number(card);
-							if (typeof number != "number" || number <= numbers[0] || number >= numbers[1]) return num + 10;
-						},
-					},
-				})
-				.filter((event, player) => {
-					return event._tongwei_checked;
-				})
-				.then(() => {
-					var number = get.number(trigger.card);
-					var numbers = get.info(event.name).numbers;
-					event.playerx = get.info(event.name).playerx;
-					if (typeof number != "number" || number <= numbers[0] || number >= numbers[1]) event.finish();
-				})
-				.then(() => {
-					var playerx = event.playerx;
-					var names = ["sha", "guohe"].filter(name => playerx.canUse({ name: name, isCard: true }, player, false));
-					if (!names.length) event.finish();
-					else if (names.length == 1) event._result = { links: [[null, null, names[0]]] };
-					else
-						playerx
-							.chooseButton([`请选择要视为对${get.translation(player)}使用的牌`, [names, "vcard"]], true)
-							.set("ai", button => {
-								return button.link[0][2] == _status.event.choice;
-							})
-							.set(
-								"choice",
-								(function () {
-									var list = names
-										.map(name => {
-											return [name, get.effect(player, { name: name, isCard: true }, playerx, playerx)];
-										})
-										.sort((a, b) => {
-											return b[1] - a[1];
-										});
-									return list[0][0];
-								})()
-							);
-				})
-				.then(() => {
-					var name = result.links[0][2];
-					var card = { name: name, isCard: true },
-						playerx = event.playerx;
-					if (playerx.canUse(card, player, false)) playerx.useCard(card, player, "tongwei");
-				});
-		},
-		ai: {
-			expose: 0.2,
-			order: 7,
-			threaten: 2.2,
-			result: {
-				target: -1,
-			},
-		},
-	},
-	cuguo: {
-		audio: 2,
-		trigger: { player: ["shaMiss", "eventNeutralized"] },
-		filter: function (event, player) {
-			if (event.type != "card" && event.name != "_wuxie") return false;
-			if (
-				!event.target ||
-				!event.target.isIn() ||
-				!player.canUse(
-					{
-						name: event.card.name,
-						nature: event.card.nature,
-						isCard: true,
-					},
-					event.target,
-					false
-				)
-			)
-				return false;
-			if (!player.hasCard(card => lib.filter.cardDiscardable(card, player), "he")) return false;
-			var history = game.getGlobalHistory("everything");
-			for (var evt of history) {
-				if (evt._neutralized || (evt.responded && (!evt.result || !evt.result.bool))) {
-					var evtx = evt.getParent();
-					return evtx.name == "useCard" && evtx.player == player && evt == event;
-				}
-			}
-			return false;
-		},
-		forced: true,
-		direct: true,
-		content: function () {
-			"step 0";
-			var card = {
-				name: trigger.card.name,
-				nature: trigger.card.nature,
-				isCard: true,
-				storage: { cuguo: true },
-			};
-			event.card = card;
-			player.chooseToDiscard("蹙国：请弃置一张牌", `视为你对${get.translation(trigger.target)}使用一张${get.translation(card.nature || "")}【${get.translation(card.name)}】`, "he", true).set("logSkill", ["cuguo", trigger.target]);
-			"step 1";
-			if (player.canUse(card, trigger.target, false)) {
-				player.useCard(card, trigger.target);
-				player
-					.when("useCardAfter")
-					.filter(event => {
-						return event.card.storage && event.card.storage.cuguo;
-					})
-					.then(() => {
-						if (
-							game.hasGlobalHistory("everything", evt => {
-								if (evt._neutralized || (evt.responded && (!evt.result || !evt.result.bool))) {
-									if (evt.getParent() == trigger) return true;
-								}
-								return false;
-							})
-						) {
-							player.loseHp();
-						}
-					});
-			}
-		},
-	},
-	//陈式
-	qingbei: {
-		audio: 2,
-		trigger: { global: "roundStart" },
-		direct: true,
-		content: function () {
-			"step 0";
-			var next = player.chooseButton(['###擎北：是否选择任意种花色？###<div class="text center">你不能于本轮能使用这些花色，且使用牌后摸等同于选择花色数的牌</div>', [lib.suit.map(i => ["", "", "lukai_" + i]), "vcard"]], [1, 4]);
-			next.set("ai", button => {
-				var player = _status.event.player;
-				var suit = button.link[2].slice(6);
-				var val = player
-					.getCards("hs", { suit: suit })
-					.map(card => {
-						return get.value(card) + player.getUseValue(card) / 3;
-					})
-					.reduce((p, c) => {
-						return p + c;
-					}, 0);
-				if (val > 10 && ui.selected.buttons.length > 0) return -1;
-				if (val > 6 && ui.selected.buttons.length == 2) return -1;
-				if (ui.selected.buttons.length == 3) return -1;
-				return 1 + 1 / val;
-			});
-			"step 1";
-			if (result.bool) {
-				var suits = result.links.map(i => i[2].slice(6));
-				player.logSkill("qingbei");
-				player.addTempSkill("qingbei_effect", "roundStart");
-				player.setStorage("qingbei_effect", suits);
-				player.markSkill("qingbei_effect");
-			}
-		},
-		ai: {
-			threaten: 2.3,
-		},
-		subSkill: {
-			effect: {
-				audio: "qingbei",
-				trigger: { player: "useCardAfter" },
-				charlotte: true,
-				onremove: true,
-				forced: true,
-				filter: function (event, player) {
-					if (!lib.suit.includes(get.suit(event.card))) return false;
-					return player.getStorage("qingbei_effect").length;
-				},
-				content: function () {
-					player.draw(player.getStorage("qingbei_effect").length);
-				},
-				mark: true,
-				intro: {
-					content: storage => `本轮内不能使用${get.translation(storage)}花色的牌，且使用一张有花色的牌后摸${get.cnNumber(storage.length)}张牌`,
-				},
-				mod: {
-					cardEnabled: function (card, player) {
-						if (player.getStorage("qingbei_effect").includes(get.suit(card))) return false;
-					},
-					cardSavable: function (card, player) {
-						if (player.getStorage("qingbei_effect").includes(get.suit(card))) return false;
-					},
-				},
-			},
-		},
-	},
-	//费曜
-	zhenfeng: {
-		audio: 2,
-		trigger: { global: "useCard" },
-		usable: 1,
-		filter: function (event, player) {
-			return event.player != player && event.player == _status.currentPhase && event.player.countCards("h") <= event.player.getHp();
-		},
-		check: function (event, player) {
-			var type = get.type2(event.card, event.player);
-			if (type == "equip" && event.player.hasCard(card => event.player.hasValueTarget(card))) return false;
-			if (get.attitude(player, event.player) > 0 && event.player.getHp() + event.player.countCards("hs", ["shan", "caochuan"]) <= 3) return false;
-			return true;
-		},
-		onremove: true,
-		logTarget: "player",
-		content: function () {
-			"step 0";
-			var choices = Array.from({ length: trigger.player.countCards("h") + 1 }).map((_, i) => get.cnNumber(i, true));
-			var type = get.type2(trigger.card, trigger.player);
-			player
-				.chooseControl(choices)
-				.set("prompt", "镇锋：猜测其手牌中的" + get.translation(type) + "牌数")
-				.set("ai", () => {
-					return _status.event.choice;
-				})
-				.set(
-					"choice",
-					(function () {
-						var num = trigger.player.countCards("h", card => get.type2(card) == type);
-						var knownNum = trigger.player.countKnownCards(_status.event.player, card => get.type2(card) == type);
-						if (trigger.player.isAllCardsKnown(_status.event.player)) {
-							return knownNum;
-						}
-						var restNum = num - knownNum;
-						var numx;
-						if (type == "basic") numx = num + Math.floor(Math.random() * restNum + 1);
-						else if (type == "trick") {
-							if (num > 2) numx = 2;
-							else numx = 1;
-							if (Math.random() < 0.5) {
-								numx += Math.random() > 0.5 ? 1 : -1;
-							}
-						} else {
-							numx = [0, 1].randomGet();
-						}
-						if (numx < knownNum) numx = knownNum;
-						else if (numx >= choices.length) numx = choices.length - 1;
-						return numx;
-					})()
-				);
-			"step 1";
-			var guessedNum = result.index;
-			player.chat("我猜" + get.cnNumber(guessedNum) + "张");
-			game.log(player, "猜测", trigger.player, "有", get.cnNumber(guessedNum) + "张" + get.translation(type) + "牌");
-			event.guessedNum = guessedNum;
-			game.delay();
-			"step 2";
-			var type = get.type2(trigger.card, trigger.player);
-			var count = trigger.player.countCards("h", card => get.type2(card) == type);
-			var guessedNum = event.guessedNum;
-			if (count == guessedNum) {
-				player.popup("洗具");
-				game.log(player, "猜测", "#g正确");
-				if (player.countMark("zhenfeng") < 5) player.addMark("zhenfeng", 1, false);
-				player.draw(player.countMark("zhenfeng"));
-				if (player.canUse("sha", trigger.player, false)) player.useCard({ name: "sha", isCard: true }, trigger.player);
-			} else {
-				player.popup("杯具");
-				game.log(player, "猜测", "#y错误");
-				player.clearMark("zhenfeng");
-				if (Math.abs(count - guessedNum) > 1 && trigger.player.canUse("sha", player, false)) {
-					trigger.player.useCard({ name: "sha", isCard: true }, player, false, "noai");
-				}
-			}
-		},
-		intro: {
-			content: "已连续猜对#次",
-		},
-	},
-	//新杀小加强 李严
-	dcduliang: {
-		enable: "phaseUse",
-		usable: 1,
-		filterTarget: function (card, player, target) {
-			return player != target && target.countCards("he") > 0;
-		},
-		audio: "duliang",
-		content: function () {
-			"step 0";
-			player.gainPlayerCard(target, "he", true);
-			"step 1";
-			var name = get.translation(target);
-			player
-				.chooseControl(function () {
-					return Math.random() < 0.5 ? "选项一" : "选项二";
-				})
-				.set("prompt", "督粮：请选择一项")
-				.set("choiceList", ["你观看牌堆顶的两张牌，然后令" + name + "获得其中的一或两张基本牌", "令" + name + "于下回合的摸牌阶段额外摸一张牌"]);
-			"step 2";
-			if (result.control == "选项一") {
-				var cards = get.cards(2),
-					bool = false;
-				event.cards = cards;
-				game.cardsGotoOrdering(cards);
-				for (var card of cards) {
-					if (get.type(card) == "basic") {
-						bool = true;
-						break;
-					}
-				}
-				player
-					.chooseButton(["督粮：选择令" + get.translation(target) + "获得的牌", cards], [1, 2], bool)
-					.set("filterButton", button => {
-						return get.type(button.link) == "basic";
-					})
-					.set("ai", button => {
-						return _status.event.sgn * get.value(button.link);
-					})
-					.set("sgn", get.sgnAttitude(player, target) > 0);
-			} else {
-				target.addTempSkill("dcduliang2", { player: "phaseAfter" });
-				target.addMark("dcduliang2", 1, false);
-				event.finish();
-			}
-			"step 3";
-			if (result.bool) {
-				var cardsx = result.links;
-				target.gain(cardsx, "draw");
-				game.log(target, "获得了" + get.cnNumber(cardsx.length) + "张牌");
-				cards.removeArray(cardsx);
-				cards.reverse();
-			}
-			for (var i = 0; i < cards.length; i++) {
-				ui.cardPile.insertBefore(cards[i], ui.cardPile.firstChild);
-			}
-			game.updateRoundNumber();
-		},
-		ai: {
-			order: 4,
-			result: {
-				target: -1,
-				player: 0.1,
-			},
-		},
-	},
-	dcduliang2: {
-		trigger: { player: "phaseDrawBegin" },
-		forced: true,
-		mark: true,
-		audio: false,
-		onremove: true,
-		charlotte: true,
-		intro: {
-			content: "下回合的摸牌阶段额外摸#张牌",
-		},
-		content: function () {
-			trigger.num += player.countMark("dcduliang2");
-		},
-	},
-	//苏飞
-	shuojian: {
-		audio: 2,
-		enable: "phaseUse",
-		usable: 3,
-		filterTarget: lib.filter.notMe,
-		filterCard: true,
-		position: "he",
-		discard: false,
-		lose: false,
-		delay: false,
-		check: function (card, player) {
-			return 6 - get.value(card);
-		},
-		content: function () {
-			"step 0";
-			player.give(cards, target);
-			"step 1";
-			var num = 3 - get.skillCount("shuojian") + 1;
-			event.num = num;
-			event.num2 = num;
-			if (event.num == 0) event.finish();
-			"step 2";
-			var forced = num != event.num2;
-			var prompt = "###" + get.translation(player) + "对你发动了【数谏】###视为使用" + get.cnNumber(num) + "张【过河拆桥】" + (forced ? "" : "且" + get.translation(player) + "此技能本回合失效，或点击“取消”令其摸" + get.cnNumber(num) + "张牌");
-			if (!target.hasUseTarget({ name: "guohe" })) event._result = { bool: false };
-			else
-				target
-					.chooseUseTarget(prompt, "guohe", forced)
-					.set("ai", function () {
-						var evt = _status.event;
-						if (evt.name == "chooseTarget") evt = evt.getParent();
-						if (!evt.goon) return 0;
-						return get.effect_use.apply(this, arguments);
-					})
-					.set("goon", target.getUseValue({ name: "guohe" }) > (get.sgnAttitude(target, player) * player.getUseValue({ name: "wuzhong" })) / (2 - num * 0.4));
-			"step 3";
-			if (!result.bool) {
-				player.draw(num);
-				if (num > 1) player.chooseToDiscard("he", num - 1, true);
-				event.finish();
-			}
-			"step 4";
-			if (--event.num2 > 0) {
-				event.goto(2);
-			} else player.tempBanSkill("shuojian");
-		},
-		ai: {
-			expose: 0.15,
-			order: 8,
-			result: { target: 1 },
-		},
-	},
-	//谯周
-	shiming: {
-		audio: 2,
-		trigger: { global: "phaseDrawBegin1" },
-		filter: function (event, player) {
-			return !player.hasSkill("shiming_round");
-		},
-		check: function (event, player) {
-			return true; //get.attitude(player,event.player)<0||get.damageEffect(event.player,event.player,player)>0;
-		},
-		logTarget: "player",
-		content: function () {
-			"step 0";
-			player.addTempSkill("shiming_round", "roundStart");
-			var cards = get.cards(3);
-			player
-				.chooseButton(["识命：是否将其中一张置于牌堆底？", cards.slice(0)])
-				.set("ai", button => {
-					var att = _status.event.att,
-						damage = _status.event.damage,
-						val = get.value(button.link, _status.event.player);
-					if ((att > 0 && damage < 0) || (att <= 0 && damage > 0)) return 6 - val;
-					return val - 5.99;
-				})
-				.set("att", get.attitude(player, trigger.player))
-				.set("damage", get.damageEffect(trigger.player, trigger.player, player) > 0 && trigger.player.hp <= 3 ? 1 : -1);
-			while (cards.length) ui.cardPile.insertBefore(cards.pop(), ui.cardPile.firstChild);
-			"step 1";
-			if (result.bool) {
-				var card = result.links[0];
-				card.fix();
-				ui.cardPile.appendChild(card);
-				player.popup("一下", "wood");
-				game.log(player, "将一张牌置于了牌堆底");
-			}
-			"step 2";
-			trigger.player
-				.chooseBool("是否跳过摸牌阶段并对自己造成1点伤害，然后从牌堆底摸三张牌？")
-				.set("ai", () => _status.event.bool)
-				.set("bool", get.damageEffect(trigger.player, trigger.player) >= -6 || trigger.player.hp > 3);
-			"step 3";
-			if (result.bool) {
-				trigger.cancel();
-				trigger.player.damage(trigger.player);
-			} else event.finish();
-			"step 4";
-			trigger.player.draw(3, "bottom");
-		},
-		subSkill: {
-			round: {
-				mark: true,
-				intro: { content: "本轮已发动〖识命〗" },
-			},
-		},
-	},
-	jiangxi: {
-		audio: 2,
-		trigger: { global: "phaseEnd" },
-		filter: function (event, player) {
-			var zhu = game.findPlayer(i => i.getSeatNum() == 1);
-			return (
-				(zhu &&
-					player.hasSkill("shiming_round") &&
-					(game.getGlobalHistory("changeHp", evt => {
-						return evt.player == zhu && evt._dyinged;
-					}).length > 0 ||
-						zhu.getHistory("damage").length == 0)) ||
-				!game.hasPlayer2(current => current.getHistory("damage").length > 0)
-			);
-		},
-		direct: true,
-		content: function () {
-			"step 0";
-			if (player.hasSkill("shiming_round")) {
-				var zhu = game.findPlayer(i => i.getSeatNum() == 1);
-				if (
-					game.getGlobalHistory("changeHp", evt => {
-						return evt.player == zhu && evt._dyinged;
-					}).length > 0 ||
-					zhu.getHistory("damage").length == 0
-				) {
-					player.chooseBool(get.prompt("jiangxi"), "重置〖识命〗");
-				}
-			} else event.goto(2);
-			"step 1";
-			if (result.bool) {
-				player.logSkill("jiangxi");
-				event.logged = true;
-				player.removeSkill("shiming_round");
-				player.draw();
-			}
-			"step 2";
-			if (!game.hasPlayer2(current => current.getHistory("damage").length > 0)) {
-				player
-					.chooseBool(get.prompt("jiangxi"), "与" + get.translation(trigger.player) + "各摸一张牌")
-					.set("ai", () => _status.event.bool)
-					.set("bool", trigger.player.getUseValue({ name: "wuzhong" }) + player.getUseValue({ name: "wuzhong" }) > 0);
-			} else event.finish();
-			"step 3";
-			if (result.bool) {
-				if (!event.logged) player.logSkill("jiangxi");
-				trigger.player.draw("nodelay");
-				player.draw();
-			}
-		},
-	},
-	//韩龙
-	duwang: {
-		audio: 2,
-		trigger: {
-			global: "phaseBefore",
-			player: "enterGame",
-		},
-		forced: true,
-		filter: function (event, player) {
-			return event.name != "phase" || game.phaseNumber == 0;
-		},
-		content: function () {
-			var cards = [];
-			for (var i = 0; i < ui.cardPile.childNodes.length; i++) {
-				var card = ui.cardPile.childNodes[i];
-				if (card.name != "sha") cards.push(card);
-				if (cards.length >= 5) break;
-			}
-			if (cards.length) player.addToExpansion(cards, "gain2").gaintag.add("duwang");
-		},
-		marktext: "刺",
-		intro: {
-			name: "刺",
-			name2: "刺",
-			content: "expansion",
-			markcount: "expansion",
-		},
-		mod: {
-			globalFrom: function (from, to, distance) {
-				return distance + Math.min(1, from.getExpansions("duwang").filter(i => i.name != "sha").length);
-			},
-			globalTo: function (from, to, distance) {
-				return distance + Math.min(1, to.getExpansions("duwang").filter(i => i.name != "sha").length);
-			},
-		},
-	},
-	cibei: {
-		audio: 2,
-		trigger: { global: "cardsDiscardAfter" },
-		filter: function (event, player) {
-			if (!player.getExpansions("duwang").filter(i => i.name != "sha").length) return false;
-			var evt = event.getParent();
-			if (evt.name != "orderingDiscard") return false;
-			var evtx = evt.relatedEvent || evt.getParent();
-			return (
-				evtx.name == "useCard" &&
-				evtx.card.name == "sha" &&
-				event.cards.filterInD("d").length &&
-				game.hasPlayer2(current =>
-					current.hasHistory("sourceDamage", evtxx => {
-						return evtxx.card == evtx.card;
-					})
-				)
-			);
-		},
-		direct: true,
-		group: "cibei_fullyReady",
-		content: function () {
-			"step 0";
-			player.chooseButton(["###" + get.prompt("cibei") + '###<div class="text center">将一张“刺”置入弃牌堆，并将' + get.translation(trigger.cards.filterInD("d")) + "置入“刺”</div>", player.getExpansions("duwang")]).set("filterButton", button => {
-				return button.link.name != "sha";
-			});
-			"step 1";
-			if (result.bool) {
-				player.logSkill("cibei");
-				player.loseToDiscardpile(result.links);
-				player.addToExpansion(trigger.cards.filterInD("d"), "gain2").gaintag.add("duwang");
-				if (game.hasPlayer(current => current.countDiscardableCards(player, "hej") > 0))
-					player
-						.chooseTarget("刺北：弃置一名角色区域内的一张牌", true, (card, player, target) => {
-							return target.countDiscardableCards(player, "hej") > 0;
-						})
-						.set("ai", target => {
-							return get.effect(target, { name: "guohe" }, _status.event.player);
-						});
-			} else event.finish();
-			"step 2";
-			if (result.bool) {
-				var target = result.targets[0];
-				player.line(target);
-				player.discardPlayerCard(target, "hej", true);
-				player.addExpose(0.1);
-			}
-		},
-		ai: {
-			combo: "duwang",
-		},
-		subSkill: {
-			fullyReady: {
-				trigger: { global: "phaseEnd" },
-				forced: true,
-				locked: false,
-				filter: function (event, player) {
-					var storage = player.getExpansions("duwang");
-					return storage.length > 0 && storage.every(i => i.name == "sha");
-				},
-				content: function () {
-					player.gain(player.getExpansions("duwang"), "gain2").gaintag.add("cibei_mark");
-					player.addSkill("cibei_mark");
-				},
-			},
-			mark: {
-				trigger: { player: "useCard1" },
-				onremove: true,
-				charlotte: true,
-				silent: true,
-				firstDo: true,
-				filter: function (event, player) {
-					return player.hasHistory("lose", evt => {
-						if (evt.getParent() != event) return false;
-						for (var i in evt.gaintag_map) {
-							if (evt.gaintag_map[i].includes("cibei_mark")) return true;
-						}
-						return false;
-					});
-				},
-				content: function () {
-					if (trigger.addCount !== false) {
-						trigger.addCount = false;
-						var stat = player.getStat().card,
-							name = trigger.card.name;
-						if (typeof stat[name] == "number") stat[name]--;
-					}
-				},
-				mod: {
-					ignoredHandcard: function (card, player) {
-						if (card.hasGaintag("cibei_mark")) return true;
-					},
-					cardDiscardable: function (card, player, name) {
-						if (card.hasGaintag("cibei_mark")) return false;
-					},
-					canBeDiscarded: function (card) {
-						if (card.hasGaintag("cibei_mark")) return false;
-					},
-					targetInRange: function (card, player, target) {
-						if (!card.cards) return;
-						for (var i of card.cards) {
-							if (i.hasGaintag("cibei_mark")) return true;
-						}
-					},
-					cardUsable: function (card, player) {
-						if (!card.cards) return;
-						for (var i of card.cards) {
-							if (i.hasGaintag("cibei_mark")) return true;
-						}
-					},
-				},
-			},
-		},
-	},
-	//武安国
-	diezhang: {
-		audio: 2,
-		locked: false,
-		zhuanhuanji: function (player, skill) {
-			if (!player.storage.duanwan) player.storage[skill] = !player.storage[skill];
-		},
-		trigger: { global: ["eventNeutralized", "shaMiss"] },
-		filter: function (event, player) {
-			if (player.hasSkill("diezhang_used")) return false;
-			if (event.type != "card") return false;
-			var evt = event._neutralize_event;
-			var user, responder;
-			if (event.name == "sha") {
-				user = event.player;
-				responder = event.target;
-			} else {
-				if (evt.type != "card") return false;
-				user = event.player;
-				responder = evt.player;
-			}
-			if (!player.storage.diezhang) {
-				if (user != player || responder == player) return false;
-				return player.countDiscardableCards(player, "he") > 0 && player.canUse("sha", responder, false);
-			} else {
-				if (user == player || responder != player) return false;
-				return player.canUse("sha", user, false);
-			}
-		},
-		direct: true,
-		content: function () {
-			"step 0";
-			var evt = trigger._neutralize_event;
-			var user, responder;
-			if (trigger.name == "sha") {
-				user = trigger.player;
-				responder = trigger.target;
-			} else {
-				user = trigger.player;
-				responder = evt.player;
-			}
-			var num = player.storage.duanwan ? 2 : 1;
-			event.num = num;
-			if (!player.storage.diezhang) {
-				var target = responder;
-				event.target = target;
-				var next = player
-					.chooseToDiscard(get.prompt("diezhang", target), "弃置一张牌，视为对其使用" + get.cnNumber(num) + "张【杀】", "he")
-					.set("ai", card => {
-						if (_status.event.goon) return 6 - get.value(card);
-						return 0;
-					})
-					.set("logSkill", ["diezhang", target]);
-			} else {
-				var target = user;
-				event.target = target;
-				var next = player.chooseBool(get.prompt("diezhang", target), "摸" + get.cnNumber(num) + "张牌，视为对其使用一张【杀】").set("ai", () => _status.event.goon);
-			}
-			next.set("goon", get.effect(target, { name: "sha" }, player, player) > 0);
-			"step 1";
-			if (result.bool) {
-				if (player.storage.duanwan) player.addTempSkill("diezhang_used");
-				player.changeZhuanhuanji("diezhang");
-				if (!result.cards || !result.cards.length) {
-					player.logSkill("diezhang", target);
-					player.draw(num, "nodelay");
-					player.useCard({ name: "sha", isCard: true }, target, false);
-				} else while (num--) player.useCard({ name: "sha", isCard: true }, target, false);
-			}
-		},
-		marktext: "☯",
-		mark: true,
-		intro: {
-			content: function (storage, player) {
-				var cnNum = get.cnNumber(player.storage.duanwan ? 2 : 1);
-				if (storage) return "当其他角色使用牌被你抵消后，你可以摸" + cnNum + "张牌，视为对其使用一张【杀】。";
-				return "当你使用牌被其他角色抵消后，你可以弃置一张牌，视为对其使用" + cnNum + "张【杀】。";
-			},
-		},
-		mod: {
-			cardUsable: function (card, player, num) {
-				if (!player.storage.duanwan && card.name == "sha") return num + 1;
-			},
-		},
-		subSkill: { used: { charlotte: true } },
-	},
-	duanwan: {
-		audio: 2,
-		enable: "chooseToUse",
-		skillAnimation: true,
-		animationColor: "soil",
-		limited: true,
-		filter: function (event, player) {
-			return event.type == "dying" && player == event.dying;
-		},
-		content: function () {
-			player.changeZhuanhuanji("diezhang");
-			player.awakenSkill("duanwan");
-			var num = 2 - player.hp;
-			if (num > 0) player.recover(num);
-		},
-		ai: {
-			save: true,
-			skillTagFilter: function (player, tag, target) {
-				return player == target;
-			},
-			result: { player: 1 },
-		},
-	},
-	//李婉
-	liandui: {
-		audio: 2,
-		trigger: { global: "useCard" },
-		filter: function (event, player) {
-			const history = game.getAllGlobalHistory("useCard");
-			const index = history.indexOf(event);
-			if (index <= 0) return false;
-			const previous = history[index - 1].player;
-			if (event.player == player && previous != player && previous.isIn()) return true;
-			if (event.player != player && previous == player) return true;
-			return false;
-		},
-		async cost(event, trigger, player) {
-			if (!trigger.player) return;
-			const history = game.getAllGlobalHistory("useCard");
-			const index = history.indexOf(trigger);
-			const previous = history[index - 1].player;
-			const { result } = await trigger.player
-				.chooseBool("是否对" + get.translation(previous) + "发动【联对】？", "令" + get.translation(previous) + "摸两张牌")
-				.set("ai", () => _status.event.bool)
-				.set("bool", get.effect(previous, { name: "draw" }, trigger.player, trigger.player) > 0);
-			if (result.bool) event.result = { bool: true, cost_data: previous };
-		},
-		async content(event, trigger, player) {
-			const { cost_data: previous } = event;
-			previous.draw(2);
-		},
-	},
-	biejun: {
-		audio: 2,
-		global: "biejun_give",
-		trigger: { player: "damageBegin4" },
-		filter: function (event, player) {
-			return (
-				!player.hasSkill("biejun_used") &&
-				player.countCards("h", card => {
-					return card.hasGaintag("biejun");
-				}) == 0
-			);
-		},
-		prompt2: "翻面并防止此伤害",
-		check: function (event, player) {
-			return player.isTurnedOver() || event.num >= player.hp || get.distance(_status.currentPhase, player, "absolute") >= 3;
-		},
-		content: function () {
-			player.addTempSkill("biejun_used");
-			player.turnOver();
-			trigger.cancel();
-		},
-		ai: {
-			effect: {
-				target: function (card, player, target) {
-					if (player.hasSkillTag("jueqing", false, target)) return [1, -2];
-					if (get.tag(card, "damage")) {
-						if (player.getNext() == target && lib.skill.biejun.filter(null, target) && target.isTurnedOver()) return [0, 1];
-					}
-				},
-			},
-		},
-		subSkill: {
-			used: { charlotte: true },
-			give: {
-				audio: 2,
-				enable: "phaseUse",
-				usable: 1,
-				filter: function (event, player) {
-					if (!player.countCards("h")) return false;
-					var targets = game.filterPlayer(function (current) {
-						return current != player && current.hasSkill("biejun");
-					});
-					if (!targets.length) return false;
-					return true;
-				},
-				selectCard: 1,
-				filterCard: true,
-				filterTarget: function (card, player, target) {
-					return target.hasSkill("biejun");
-				},
-				selectTarget: function () {
-					var player = _status.event.player;
-					var targets = game.filterPlayer(function (current) {
-						return current != player && current.hasSkill("biejun");
-					});
-					return targets.length > 1 ? 1 : -1;
-				},
-				complexSelect: true,
-				prompt: function () {
-					var player = _status.event.player;
-					var targets = game.filterPlayer(function (current) {
-						return current != player && current.hasSkill("biejun");
-					});
-					return "将一张手牌交给" + get.translation(targets) + (targets.length > 1 ? "中的一人" : "");
-				},
-				position: "h",
-				discard: false,
-				lose: false,
-				delay: false,
-				check: function (card) {
-					var player = _status.event.player;
-					if (
-						game.hasPlayer(function (current) {
-							return lib.skill.biejun_give.filterTarget(null, player, current) && get.attitude(player, current) > 0;
-						})
-					) {
-						return 5 - get.value(card);
-					}
-					return -get.value(card);
-				},
-				content: function () {
-					game.trySkillAudio("biejun", target);
-					player.give(cards, target).gaintag.add("biejun");
-					target.addTempSkill("biejun_tag");
-				},
-				ai: {
-					order: 2,
-					result: { target: 1 },
-				},
-			},
-			tag: {
-				charlotte: true,
-				forced: true,
-				onremove: function (player) {
-					player.removeGaintag("biejun");
-				},
-			},
-		},
-	},
-	//诸葛尚
-	sangu: {
-		audio: 2,
-		trigger: { player: "phaseJieshuBegin" },
-		direct: true,
-		getEffect: function (player, target, event, list1, list2) {
-			let att = get.attitude(player, target);
-			if (att === 0) return 0;
-			let getv = function (name, player, arg) {
-				let v = event.getTempCache("sangu", player.id + name);
-				if (typeof v === "number") return v;
-				v = player.getUseValue({ name: name, storage: { sangu: true } }, arg);
-				event.putTempCache("sangu", player.id + name, v);
-				return v;
-			};
-			if (att < 0) {
-				for (let i of list1) {
-					if (getv(i, target) <= 0 || getv(i, target) <= 0) return -att * Math.sqrt(get.threaten(target)) * 2;
-				}
-				return 0;
-			} else {
-				let list = list1.concat(player.hp > 1 ? list2 : []),
-					eff = 0;
-				list.sort(function (a, b) {
-					return getv(b, target) - getv(a, target);
-				});
-				list = list.slice(3);
-				for (let i of list) {
-					let res = getv(i, target);
-					if (res <= 5) break;
-					else eff += res;
-				}
-				return Math.sqrt(eff / 1.5) * att;
-			}
-		},
-		content: function () {
-			"step 0";
-			event.list1 = [];
-			event.list2 = [];
-			event.used = [];
-			player.getHistory("useCard", function (evt) {
-				event.used.add(evt.card.name);
-			});
-			for (let name of lib.inpile) {
-				let add = false,
-					type = get.type(name);
-				if (name === "sha") add = true;
-				else if (type === "trick") {
-					let info = lib.card[name];
-					if (info && !info.singleCard && !info.notarget) add = true;
-				}
-				if (!add) continue;
-				if (event.used.includes(name)) event.list1.push(name);
-				else event.list2.push(name);
-			}
-			if (!event.list1.length && !event.list2.length) event.finish();
-			else
-				player
-					.chooseTarget(get.prompt2("sangu"), lib.filter.notMe)
-					.set("ai", function (target) {
-						return lib.skill.sangu.getEffect(_status.event.player, target, _status.event.getTrigger(), _status.event.list1, _status.event.list2);
-					})
-					.set("list1", event.list1)
-					.set("list2", event.list2);
-			"step 1";
-			if (result.bool) {
-				var target = result.targets[0];
-				player.logSkill("sangu", target);
-				event.target = target;
-			} else event.finish();
-			"step 2";
-			var dialog = ["为" + get.translation(target) + "选择至多三个牌名"];
-			if (event.list1.length) {
-				dialog.push('<div class="text center">本回合已使用过的牌</div>');
-				dialog.push([event.list1.map(i => [get.type(i), "", i]), "vcard"]);
-			}
-			if (event.list2.length) {
-				dialog.push('<div class="text center">本回合未使用过的牌</div>');
-				dialog.push([event.list2.map(i => [get.type(i), "", i]), "vcard"]);
-			}
-			player
-				.chooseButton(dialog, true, [1, 3])
-				.set("ai", function (button) {
-					let name = button.link[2],
-						list = _status.event.list,
-						player = _status.event.player,
-						target = _status.event.getParent().target,
-						trigger = _status.event.getTrigger(),
-						getv = (name, player) => {
-							let v = trigger.getTempCache("sangu", player.id + name);
-							if (typeof v === "number") return v;
-							v = player.getUseValue({ name: name, storage: { sangu: true } });
-							trigger.putTempCache("sangu", player.id + name, v);
-							return v;
-						};
-					if (get.attitude(player, target) < 0) {
-						if (!list.includes(name)) return 0;
-						return -getv(name, target);
-					} else {
-						if (player.hp < 2 && !list.includes(name)) return 0;
-						let val = getv(name, target),
-							base = 5;
-						val = Math.min(15, val - base);
-						if (name === "wuzhong" || name === "dongzhuxianji") val += 15;
-						else if (name === "shunshou") val += 6;
-						return val;
-					}
-				})
-				.set("list", event.list1);
-			"step 3";
-			if (result.bool) {
-				var names = result.links.map(i => i[2]);
-				if (!target.storage.sangu_effect) target.storage.sangu_effect = [];
-				target.storage.sangu_effect = target.storage.sangu_effect.concat(names);
-				game.log(player, "为", target, "选择了", "#y" + get.translation(names));
-				target.addTempSkill("sangu_effect", { player: "phaseUseAfter" });
-				target.markSkill("sangu_effect");
-				var bool = true;
-				for (var i of names) {
-					if (!event.used.includes(i)) {
-						bool = false;
-						break;
-					}
-				}
-				if (bool) {
-					target.addTempSkill("sangu_prevent", { player: "phaseUseAfter" });
-					target.markAuto("sangu_prevent", [player]);
-				}
-			}
-		},
-		ai: {
-			effect: {
-				target: function (card, player, target, current) {
-					if (get.tag(card, "damage") && card.storage && card.storage.sangu) return "zeroplayertarget";
-				},
-			},
-		},
-		subSkill: {
-			effect: {
-				trigger: { player: "phaseUseBegin" },
-				charlotte: true,
-				forced: true,
-				popup: false,
-				content: function () {
-					player.addTempSkill("sangu_viewas");
-				},
-				onremove: true,
-				intro: {
-					mark: function (dialog, storage, player) {
-						if (!storage || !storage.length) return "当前无可用牌";
-						dialog.add([[storage[0]], "vcard"]);
-						if (storage.length > 1) dialog.addSmall([storage.slice(1), "vcard"]);
-					},
-					content: "$",
-				},
-			},
-			viewas: {
-				mod: {
-					hiddenCard: function (player, name) {
-						var storage = player.getStorage("sangu_effect");
-						if (storage.length) return name == storage[0];
-					},
-					cardname: function (card, player) {
-						if (_status.event.name != "chooseToUse" || _status.event.skill) return;
-						var storage = player.getStorage("sangu_effect");
-						if (storage.length) return storage[0];
-					},
-					cardnature: function (card, player) {
-						if (_status.event.name != "chooseToUse" || _status.event.skill) return;
-						var storage = player.getStorage("sangu_effect");
-						if (storage.length) return false;
-					},
-				},
-				trigger: { player: ["useCard", "respond"] },
-				forced: true,
-				charlotte: true,
-				filter: function (event, player) {
-					return event.cards.length > 0 && player.getStorage("sangu_effect").length > 0;
-				},
-				content: function () {
-					if (!trigger.card.storage) trigger.card.storage = {};
-					trigger.card.storage.sangu = true;
-					player.unmarkAuto("sangu_effect", [player.getStorage("sangu_effect")[0]]);
-				},
-			},
-			prevent: {
-				trigger: { source: "damageBegin2" },
-				forced: true,
-				charlotte: true,
-				onremove: true,
-				filter: function (event, player) {
-					return event.card && event.card.storage && event.card.storage.sangu && player.getStorage("sangu_prevent").includes(event.player);
-				},
-				content: function () {
-					trigger.cancel();
-				},
-			},
-		},
-	},
-	yizu: {
-		audio: 2,
-		trigger: { target: "useCardToTargeted" },
-		forced: true,
-		usable: 1,
-		filter: function (event, player) {
-			return player.isDamaged() && player.hp <= event.player.hp && (event.card.name == "sha" || event.card.name == "juedou");
-		},
-		content: function () {
-			player.recover();
-		},
-		ai: {
-			effect: {
-				target_use(card, player, target, current) {
-					if (target.isHealthy() || (card.name != "sha" && card.name != "juedou")) return;
-					if (target.storage.counttrigger && target.storage.counttrigger.yizu && current < 0) return 5;
-					if (player.hp < target.hp) return;
-					if (current > 0) return 1.2;
-					if (get.attitude(player, target) >= 0) return;
-					var copy = get.effect(target, { name: "shacopy" }, player, player);
-					if (
-						copy > 0 &&
-						player.isPhaseUsing() &&
-						Math.min(
-							player.getCardUsable("sha"),
-							player.countCards("hs", function (card) {
-								return get.name(card) == "sha" && player.canUse(card, target, null, true);
-							})
-						) >= 2
-					)
-						return;
-					return [0, 2];
-				},
-			},
-		},
-	},
-	//轲比能
-	kousheng: {
-		audio: 2,
-		trigger: { player: "phaseUseBegin" },
-		direct: true,
-		filter: function (event, player) {
-			return player.countCards("h") > 0;
-		},
-		content: function () {
-			"step 0";
-			player
-				.chooseCard("h", [1, player.countCards("h")], get.prompt("kousheng"), "你可以选择任意张手牌，这些手牌于本回合内视为无次数限制的【杀】。但当有角色受到这些【杀】的伤害后，其可以用所有手牌交换剩余的牌。")
-				.set("standard", player.getUseValue({ name: "sha" }, null, true))
-				.set("ai", function (card) {
-					var player = _status.event.player,
-						standard = _status.event.standard;
-					if (standard <= 0) return 0;
-					var eff = player.getUseValue(card, null, true);
-					if (eff <= standard) return standard - eff + 0.1;
-					return 0;
-				});
-			"step 1";
-			if (result.bool) {
-				player.logSkill("kousheng");
-				player.addGaintag(result.cards, "kousheng");
-				player.addTempSkill("kousheng_effect");
-				game.delayx();
-			}
-		},
-		subSkill: {
-			effect: {
-				audio: "kousheng",
-				trigger: { player: "useCard1" },
-				forced: true,
-				charlotte: true,
-				firstDo: true,
-				filter: function (event, player) {
-					if (event.card.name != "sha") return false;
-					return player.hasHistory("lose", function (evt) {
-						if (evt.getParent() != event) return false;
-						for (var i in evt.gaintag_map) {
-							if (evt.gaintag_map[i].includes("kousheng")) return true;
-						}
-						return false;
-					});
-				},
-				content: function () {
-					if (!trigger.card.storage) trigger.card.storage = {};
-					trigger.card.storage.kousheng = true;
-					if (trigger.addCount !== false) {
-						trigger.addCount = false;
-						player.getStat("card").sha--;
-					}
-				},
-				onremove: function (player) {
-					player.removeGaintag("kousheng");
-				},
-				mod: {
-					cardUsable: function (card, player, target) {
-						if (card.name != "sha" || !card.cards) return;
-						for (var i of card.cards) {
-							if (i.hasGaintag("kousheng")) return Infinity;
-						}
-					},
-					cardname: function (card) {
-						if (get.itemtype(card) == "card" && card.hasGaintag("kousheng")) return "sha";
-					},
-					cardnature: function (card) {
-						if (get.itemtype(card) == "card" && card.hasGaintag("kousheng")) return false;
-					},
-				},
-				group: "kousheng_damage",
-			},
-			damage: {
-				audio: "kousheng",
-				trigger: { source: "damageSource" },
-				forced: true,
-				filter: function (event, player) {
-					if (!event.card || !event.card.storage || !event.card.storage.kousheng || event.getParent().type != "card") return false;
-					var target = event.player;
-					return (
-						target.isIn() &&
-						player.hasCard(function (card) {
-							return card.hasGaintag("kousheng");
-						}, "h")
-					);
-				},
-				content: function () {
-					"step 0";
-					var target = trigger.player;
-					event.target = target;
-					var cards = player.getCards("h", function (card) {
-						return card.hasGaintag("kousheng");
-					});
-					event.cards = cards;
-					var str = get.translation(player);
-					player.showCards(cards, str + "的【寇旌】牌");
-					if (target.countCards("h") > 0)
-						target.chooseBool("是否交换“寇旌”牌？", "用你的所有手牌交换" + str + "的下列“寇旌”牌：" + get.translation(cards)).set("ai", function () {
-							var player = _status.event.player,
-								target = _status.event.getParent().player;
-							if (player.hasShan() || player.countCards("hs", { name: ["tao", "jiu"] }) > 0 || get.attitude(player, target) >= 0) return false;
-							var hs1 = player.getCards("h"),
-								hs2 = _status.event.getParent().cards;
-							if (hs2.length >= player.hp) return true;
-							if (get.value(hs1, player) >= get.value(hs2, target)) return false;
-							return true;
-						});
-					else event.finish();
-					"step 1";
-					if (result.bool) {
-						player.swapHandcards(target, cards, target.getCards("h"));
-					}
-				},
-			},
-		},
-	},
-	//陆凯
-	lkbushi: {
-		audio: 2,
-		getBushi: function (player) {
-			if (!player.storage.lkbushi) return ["spade", "heart", "club", "diamond"];
-			return player.storage.lkbushi;
-		},
-		onremove: true,
-		trigger: { player: "phaseZhunbeiBegin" },
-		direct: true,
-		locked: false,
-		content: function () {
-			"step 0";
-			var list = lib.skill.lkbushi.getBushi(player);
-			list = list.map(function (i) {
-				return ["", "", "lukai_" + i];
-			});
-			var next = player.chooseToMove("卜筮：是否调整〖卜筮〗的花色顺序？");
-			next.set("list", [
-				[
-					"无次数限制/使用打出摸牌<br>可弃牌无效/结束阶段获得",
-					[list, "vcard"],
-					function (list) {
-						var list2 = list.map(function (i) {
-							return get.translation(i[2].slice(6));
-						});
-						return "你使用" + list2[0] + "牌时无次数限制；使用或打出" + list2[1] + "时，摸一张牌；<br>成为" + list2[2] + "牌目标后可弃一张牌无效；结束阶段获得一张" + list2[3] + "牌";
-					},
-				],
-			]);
-			next.set("processAI", function () {
-				var player = _status.event.player;
-				var list = lib.skill.lkbushi.getBushi(player);
-				var list2 = [];
-				var hs = player.getCards("hs", function (card) {
-					return player.hasValueTarget(card);
-				});
-				list.sort(function (a, b) {
-					return hs.filter(i => get.suit(i) == b).length - hs.filter(i => get.suit(i) == a).length;
-				});
-				list2.push(list.shift());
-				hs = player.getCards("hs", "sha");
-				list.sort(function (a, b) {
-					return hs.filter(i => get.suit(i) == b).length - hs.filter(i => get.suit(i) == a).length;
-				});
-				list2.unshift(list.shift());
-				list.randomSort();
-				list2.addArray(list);
-				return [list2.map(i => ["", "", "lukai_" + i])];
-			});
-			"step 1";
-			if (result.bool) {
-				var list = lib.skill.lkbushi.getBushi(player),
-					list2 = result.moved[0].map(function (i) {
-						return i[2].slice(6);
-					});
-				for (var i = 0; i < 4; i++) {
-					if (list[i] != list2[i]) {
-						player.logSkill("lkbushi");
-						player.storage.lkbushi = list2;
-						var str = "#g";
-						for (var j = 0; j < 4; j++) {
-							str += get.translation(list2[j]);
-							if (j != 3) str += "/";
-						}
-						game.log(player, "将", "#g【卜筮】", "的花色序列改为", str);
-						game.delayx();
-						break;
-					}
-				}
-			}
-		},
-		mark: true,
-		marktext: "筮",
-		intro: {
-			content: function (storage, player) {
-				var list = lib.skill.lkbushi.getBushi(player).map(i => get.translation(i));
-				return "①你使用" + list[0] + "牌无次数限制。②当你使用或打出" + list[1] + "牌后，你摸一张牌。③当你成为" + list[2] + "牌的目标后，你可以弃置一张牌，令此牌对你无效。④结束阶段开始时，你从牌堆或弃牌堆获得一张" + list[3] + "牌。⑤准备阶段开始时，你可调整此技能中四种花色的对应顺序。";
-			},
-		},
-		group: ["lkbushi_unlimit", "lkbushi_draw", "lkbushi_defend", "lkbushi_gain"],
-		subSkill: {
-			unlimit: {
-				mod: {
-					cardUsable: function (card, player) {
-						const list = lib.skill.lkbushi.getBushi(player),
-							suit = get.suit(card);
-						if (suit === "unsure" || list[0] === suit) return Infinity;
-					},
-				},
-				trigger: { player: "useCard1" },
-				forced: true,
-				popup: false,
-				silent: true,
-				firstDo: true,
-				filter: function (event, player) {
-					if (event.addCount === false) return true;
-					var list = lib.skill.lkbushi.getBushi(player);
-					return list[0] == get.suit(event.card);
-				},
-				content: function () {
-					trigger.addCount = false;
-					var stat = player.getStat().card,
-						name = trigger.card.name;
-					if (stat[name] && typeof stat[name] == "number") stat[name]--;
-				},
-			},
-			draw: {
-				audio: "lkbushi",
-				trigger: { player: ["useCard", "respond"] },
-				forced: true,
-				locked: false,
-				filter: function (event, player) {
-					var list = lib.skill.lkbushi.getBushi(player);
-					return list[1] == get.suit(event.card);
-				},
-				content: function () {
-					player.draw();
-				},
-			},
-			defend: {
-				audio: "lkbushi",
-				trigger: { target: "useCardToTargeted" },
-				direct: true,
-				filter: function (event, player) {
-					var list = lib.skill.lkbushi.getBushi(player);
-					return list[2] == get.suit(event.card) && !event.excluded.includes(player) && player.countCards("he") > 0;
-				},
-				content: function () {
-					"step 0";
-					player
-						.chooseToDiscard("he", get.prompt("lkbushi"), "弃置一张牌，令" + get.translation(trigger.card) + "对你无效")
-						.set("ai", function (card) {
-							if (_status.event.eff >= 0) return false;
-							return -_status.event.eff * 1.1 - get.value(card);
-						})
-						.set("eff", get.effect(player, trigger.card, trigger.player, player)).logSkill = ["lkbushi_defend", trigger.player];
-					"step 1";
-					if (result.bool) {
-						trigger.excluded.add(player);
-					}
-				},
-			},
-			gain: {
-				audio: "lkbushi",
-				trigger: { player: "phaseJieshuBegin" },
-				forced: true,
-				locked: false,
-				content: function () {
-					var list = lib.skill.lkbushi.getBushi(player);
-					var card = get.cardPile(function (card) {
-						return get.suit(card, false) == list[3];
-					});
-					if (card) player.gain(card, "gain2");
-				},
-			},
-		},
-	},
-	lkzhongzhuang: {
-		audio: 2,
-		trigger: { source: ["damageBegin1", "damageBegin4"] },
-		forced: true,
-		filter: function (event, player, name) {
-			if (!event.card || event.card.name != "sha" || event.getParent().type != "card") return false;
-			var range = player.getAttackRange();
-			if (name == "damageBegin1") return range > 3;
-			return range < 3 && event.num > 1;
-		},
-		content: function () {
-			if (event.triggername == "damageBegin1") trigger.num++;
-			else trigger.num = 1;
-		},
-		global: "lkzhongzhuang_ai",
-		subSkill: {
-			ai: {
-				ai: {
-					filterDamage: true,
-					skillTagFilter: function (player, tag, arg) {
-						if (arg && arg.card && arg.card.name == "sha") {
-							if (arg.player && arg.player.hasSkill("lkzhongzhuang") && arg.player.getAttackRange() < 3) return true;
-						}
-						return false;
-					},
-				},
-			},
-		},
-	},
 	//顾雍
 	olbingyi: {
 		audio: "bingyi",
