+++ character/yijiang/sort.js
@@ -6,9 +6,6 @@ const characterSort = {
 	yijiang_2015: ["caoxiu", "caorui", "zhongyao", "xiahoushi", "liuchen", "zhangyi", "zhuzhi", "quancong", "sunxiu", "gongsunyuan", "guotufengji"],
 	yijiang_2016: ["guohuanghou", "sunziliufang", "huanghao", "liyan", "sundeng", "cenhun", "zhangrang", "liuyu"],
 	yijiang_2017: ["xinxianying", "jikang", "wuxian", "qinmi", "xuezong", "xushi", "caiyong", "caojie"],
-	yijiang_2022: ["lukai", "kebineng", "zhugeshang", "liwan", "wuanguo", "hanlong", "yj_sufei", "yj_qiaozhou"],
-	yijiang_2023: ["xiahoumao", "chenshi", "sunli", "feiyao"],
-	yijiang_2024: ["linghuyu", "yj_simafu", "yj_xuangongzhu", "xukun", "xunyuxunyou", "sb_guojia"],
 };
 
 const characterSortTranslate = {
@@ -19,9 +16,6 @@ const characterSortTranslate = {
 	yijiang_2015: "一将成名2015",
 	yijiang_2016: "原创设计2016",
 	yijiang_2017: "原创设计2017",
-	yijiang_2022: "原创设计2022",
-	yijiang_2023: "一将成名2023",
-	yijiang_2024: "一将成名2024",
 };
 
 export { characterSort, characterSortTranslate };
